### 2024-06-01 #645 @pachadotdev
Capybara
Submitting Author Name: Mauricio Vargas Sepulveda
Submitting Author Github Handle: @pachadotdev
Repository:  https://github.com/pachadotdev/capybara
Version submitted:
Submission type: Standard

Version accepted: TBD

---

```
Package: capybara
Title: Fast and Memory Efficient Fitting of Linear Models With High-Dimensional
    Fixed Effects
Version: 0.5.1
    person(
        given = "Mauricio",
        family = "Vargas Sepulveda",
        email = "m.sepulveda@mail.utoronto.ca",
    )
    Formula,
    generics,
    MASS,
    stats
    fixest,
    tidyr
Depends: R(>= 3.5.0)
Description: Fast and user-friendly estimation of generalized linear models with
    multiple fixed effects and cluster the standard errors. The method to obtain
    the estimated fixed-effects coefficients is based on Stammann (2018)
    <https://arxiv.org/abs/1707.01815> and Gaure (2013)
    <https://dx.doi.org/10.1016/j.csda.2013.03.024>.
License: Apache License (>= 2)
NeedsCompilation: yes
LinkingTo: cpp11, cpp11armadillo
    pachadotdev/cpp11armadillo
    ropenscilabs/srr
Roxygen: list(markdown = TRUE, roclets = c("namespace", "rd", "srr::srr_stats_roclet"))
```

	- [x] Other:  Regression and Supervised Learning

This helps to estimate linear models with many fixed effects and implement and efficient algorithm that saves time and memory. In addition, this features a novel wrapper for Armadillo (C++).

People (mostly) in the social sciences that need multiple controls in their models. This is especially useful in Economics and International Relations.

Fixest, alpaca. This one has different design choices and a reduced number of dependencies. One attribute is that it adds multiple unit tests.

https://github.com/ropensci/software-review/issues/629
