### 2024-06-05 #645 @jooolia
Dear @pachadotdev , Thank you for your submission.

You have checked "Regression and Supervised Learning" which would entail a "Statistical software review" (a different issue template to be opened, but that is ok). Would you be willing to go through the statistical software review as described here and implementing the standards:https://stats-devguide.ropensci.org/pkgdev.html

Thanks, Julia
