### 2024-06-06 #645 @pachadotdev
> Dear @pachadotdev , Thank you for your submission.
>
> You have checked "Regression and Supervised Learning" which would entail a "Statistical software review" (a different issue template to be opened, but that is ok). Would you be willing to go through the statistical software review as described here and implementing the standards:https://stats-devguide.ropensci.org/pkgdev.html
>
> Thanks, Julia

thanks a lot !

indeed, when I opened the issue that was the only available template on my end

I already implemented https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-srr
