### 2024-06-29 #645 @jooolia
Hi @pachadotdev , sorry for the delayed response. You have the option of several different types of templates when you open an issue the statistical one is the last choice (the template is viewable here: https://github.com/ropensci/software-review/blob/main/.github/ISSUE_TEMPLATE/F-submit-statistical-software-for-review.md). Does this help? Thanks, Julia
