+++ R/cpp11.R
@@ -75,3 +75,7 @@ solve_eta2_ <- function(yadj, myadj, offset, eta) {
 sqrt_ <- function(w) {
   .Call(`_capybara_sqrt_`, w)
 }
+
+kendall_cor_ <- function(m) {
+  .Call(`_capybara_kendall_cor_`, m)
+}
