### 2024-07-04 fec825f
pass matrix by reference in kendall_cor
