+++ NAMESPACE
@@ -59,6 +59,7 @@ importFrom(rlang,":=")
 importFrom(rlang,sym)
 importFrom(stats,as.formula)
 importFrom(stats,binomial)
+importFrom(stats,complete.cases)
 importFrom(stats,gaussian)
 importFrom(stats,model.matrix)
 importFrom(stats,na.omit)
