### 2024-07-04 #645 @pachadotdev
here is the filled template

---
name: Capybara
about: Linear models with fixed effects

---

Submitting Author Name: Pacha
Version submitted: 0.2.5 (initially)
Badge grade: silver (?)
Reviewers: <TBD
---

```
Package: capybara
Title: Fast and Memory Efficient Fitting of Linear Models With High-Dimensional
    Fixed Effects
    person(
        given = "Mauricio",
        family = "Vargas Sepulveda",
        email = "m.sepulveda@mail.utoronto.ca",
    )
    Formula,
    generics,
    MASS,
    stats
    tidyr
Depends: R(>= 3.5.0)
Description: Fast and user-friendly estimation of generalized linear models with
    multiple fixed effects and cluster the standard errors. The method to obtain
    the estimated fixed-effects coefficients is based on Stammann (2018)
    <https://arxiv.org/abs/1707.01815> and Gaure (2013)
    <https://dx.doi.org/10.1016/j.csda.2013.03.024>.
License: Apache License (>= 2)
NeedsCompilation: yes
LinkingTo: cpp11, cpp11armadillo
    pachadotdev/cpp11armadillo,
    ropenscilabs/srr
Roxygen: list(markdown = TRUE, roclets = c("namespace", "rd", "srr::srr_stats_roclet"))
```

- Please indicate which of our [statistical package categories](https://stats-devguide.ropensci.org/overview.html#overview-categories) this package falls under. (Please check one or more appropriate boxes below):

- [x] A pre-submission inquiry has been approved in [issue#629](#629)

People (mostly) in the social sciences that need multiple controls in their models. This is especially useful in Economics and International Relations.

    - *An improvement on other implementations of similar algorithms in* **R**.

    Please include hyperlinked references to all other relevant software. https://github.com/amrei-stammann/alpaca/

-    What grade of badge are you aiming for? ([silver](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-badges))

* Have a demonstrated generality of usage beyond one single envisioned use case.
* Compliance with a good number of standards beyond those identified as minimally necessary.

- x] I have read the [author guide](https://devdevguide.netlify.app/authors-guide.html) and I expect to maintain this package for at least 2 years or have another maintainer identified.
