+++ src/cpp11.cpp
@@ -145,6 +145,13 @@ extern "C" SEXP _capybara_kendall_cor_(SEXP m) {
     return cpp11::as_sexp(kendall_cor_(cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<> &>>(m)));
   END_CPP11
 }
+// 05_kendall_correlation.cpp
+doubles pkendall_(doubles Q, int n);
+extern "C" SEXP _capybara_pkendall_(SEXP Q, SEXP n) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(pkendall_(cpp11::as_cpp<cpp11::decay_t<doubles>>(Q), cpp11::as_cpp<cpp11::decay_t<int>>(n)));
+  END_CPP11
+}
 
 extern "C" {
 static const R_CallMethodDef CallEntries[] = {
@@ -158,6 +165,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_capybara_group_sums_var_",      (DL_FUNC) &_capybara_group_sums_var_,      2},
     {"_capybara_inv_",                 (DL_FUNC) &_capybara_inv_,                 1},
     {"_capybara_kendall_cor_",         (DL_FUNC) &_capybara_kendall_cor_,         1},
+    {"_capybara_pkendall_",            (DL_FUNC) &_capybara_pkendall_,            2},
     {"_capybara_rank_",                (DL_FUNC) &_capybara_rank_,                1},
     {"_capybara_sandwich_",            (DL_FUNC) &_capybara_sandwich_,            2},
     {"_capybara_solve_beta_",          (DL_FUNC) &_capybara_solve_beta_,          4},
