+++ R/kendall_correlation.R
@@ -1,8 +1,132 @@
-# not exported, it is not of interest to the user and it is just for the
-# pseudo-R2 computation in the summary method
+#' @title Kendall Correlation
+#' 
+#' @description \code{\link{kendall_cor}} calculates the Kendall correlation
+#'  coefficient between two numeric vectors. It uses the algorithm described in
+#'  Knight (1966), which is based on the number of concordant and discordant
+#'  pairs. The computational complexity of the algorithm is
+#'  \eqn{O(n \log(n))}{O(n log(n))}, which is faster than the base R
+#'  implementation in \code{\link[stats]{cor}} with \code{method = "kendall"}
+#'  that has a computational complexity of \eqn{O(n^2)}{O(n^2)}. For small
+#'  vectors (i.e., less than 100 observations), the time difference is
+#'  negligible. However, for larger vectors, the difference can be substantial.
+#' 
+#'  By construction, the implementation drops missing values on a pairwise
+#'  basis. This is the same as using \code{\link[stats]{cor}} with
+#'  \code{use = "pairwise.complete.obs"}.
+#' 
+#' @param x a numeric vector.
+#' @param y a numeric vector.
+#' 
+#' @return A numeric value between -1 and 1.
+#' 
+#' @references Knight, W. R. (1966). "A Computer Method for Calculating
+#'  Kendall's Tau with Ungrouped Data". Journal of the American Statistical
+#'  Association, 61(314), 436–439.
+#' 
+#' @examples
+#' x <- c(1, 0, 2)
+#' y <- c(5, 3, 4)
+#' kendall_cor(x, y)
+#' 
+#' @export
 kendall_cor <- function(x, y) {
   arr <- cbind(x, y)
   storage.mode(arr) <- "double"
   arr <- arr[complete.cases(arr), ]
   kendall_cor_(arr)
+}
+
+#' @title Kendall Correlation Test
+#' 
+#' @description \code{\link{kendall_cor_test}} calculates p-value for the the
+#'  Kendall correlation using the exact values when the number of observations
+#'  is less than 50. For larger samples, it uses an approximation as in base R.
+#'
+#' @param x a numeric vector.
+#' @param y a numeric vector.
+#' @param alternative a character string specifying the alternative hypothesis.
+#'  The possible values are \code{"two.sided"}, \code{"greater"}, and
+#'  \code{"less"}.
+#' 
+#' @return A list with the following components:
+#' \item{statistic}{The Kendall correlation coefficient.}
+#' \item{p_value}{The p-value of the test.}
+#' \item{alternative}{A character string describing the alternative hypothesis.}
+#' 
+#' @references Knight, W. R. (1966). "A Computer Method for Calculating
+#'  Kendall's Tau with Ungrouped Data". Journal of the American Statistical
+#'  Association, 61(314), 436–439.
+#' 
+#' @examples
+#' x <- c(1, 0, 2)
+#' y <- c(5, 3, 4)
+#' kendall_cor_test(x, y)
+#' 
+#' @export
+kendall_cor_test <- function(x,y,
+  alternative = c("two.sided", "greater", "less")) {
+  alternative <- match.arg(alternative)
+
+  arr <- cbind(x, y)
+  storage.mode(arr) <- "double"
+  arr <- arr[complete.cases(arr), ]
+  r <- kendall_cor_(arr)
+  n <- nrow(arr)
+
+  if (n < 2) {
+    stop("not enough finite observations")
+  }
+  
[…]
