+++ NAMESPACE
@@ -37,6 +37,8 @@ export(fenegbin)
 export(fepoisson)
 export(fixed_effects)
 export(glance)
+export(kendall_cor)
+export(kendall_cor_test)
 export(tidy)
 importFrom(Formula,Formula)
 importFrom(MASS,negative.binomial)

[…]
