### 2024-07-05 5d08dc8
kendall cor checks
