+++ NAMESPACE
@@ -73,6 +73,7 @@ importFrom(stats,rgamma)
 importFrom(stats,rlogis)
 importFrom(stats,rnorm)
 importFrom(stats,rpois)
+importFrom(stats,sd)
 importFrom(stats,terms)
 importFrom(stats,vcov)
 importFrom(utils,combn)
