### 2024-07-06 ccf2773
streamline kendall cor functions
