+++ src/04_linear_algebra.cpp
@@ -1,0 +1,222 @@
+#include "00_main.h"
+
+// Y <- crossprod(X)
+// Y <- t(X) %*% X
+
+[[cpp11::register]] doubles_matrix<> crossprod_(const doubles_matrix<> &x,
+                                                const doubles &w,
+                                                const bool &weighted,
+                                                const bool &root_weights) {
+  Mat<double> X = as_Mat(x);
+  int P = X.n_cols;
+
+  Mat<double> res(P, P);
+
+  if (!weighted) {
+    res = X.t() * X;
+  } else {
+    Mat<double> W = as_Mat(w);
+
+    if (root_weights) {
+      W = sqrt(W);
+    }
+
+    X = X.each_col() % W;
+
+    res = X.t() * X;
+  }
+
+  return as_doubles_matrix(res);
+}
+
+// WinvJ < -solve(object[["Hessian"]] / nt.full, J)
+// Gamma < -(MX %*% WinvJ - PPsi) * v / nt.full
+// V < -crossprod(Gamma)
+
+[[cpp11::register]] doubles_matrix<>
+gamma_(const doubles_matrix<> &mx, const doubles_matrix<> &hessian,
+       const doubles_matrix<> &j, const doubles_matrix<> &ppsi,
+       const doubles &v, const SEXP &nt_full) {
+  Mat<double> MX = as_Mat(mx);
+  Mat<double> H = as_Mat(hessian);
+  Mat<double> J = as_Mat(j);
+  Mat<double> PPsi = as_Mat(ppsi);
+  Mat<double> V = as_Mat(v);
+
+  double inv_N = 1.0 / as_cpp<double>(nt_full);
+  
+  Mat<double> res = (MX * solve(H * inv_N, J)) - PPsi;
+  res = (res.each_col() % V) * inv_N;
+
+  return as_doubles_matrix(res);
+}
+
+// solve(H)
+
+[[cpp11::register]] doubles_matrix<> inv_(const doubles_matrix<> &h) {
+  Mat<double> H = inv(as_Mat(h));
+  return as_doubles_matrix(H);
+}
+
+// qr(X)$rank
+
+[[cpp11::register]] int rank_(const doubles_matrix<> &x) {
+  Mat<double> X = as_Mat(x);
+  return arma::rank(X); // SVD
+}
+
+// Beta_uncorr - solve(H / nt, B)
+
+[[cpp11::register]] doubles solve_bias_(const doubles &beta_uncorr,
+                                        const doubles_matrix<> &hessian,
+                                        const double &nt, const doubles &b) {
+  Mat<double> Beta_uncorr = as_Mat(beta_uncorr);
+  Mat<double> H = as_Mat(hessian);
+  Mat<double> B = as_Mat(b);
+
+  double inv_nt = 1.0 / nt;
+
+  return as_doubles(Beta_uncorr - solve(H * inv_nt, B));
+}
+
+// A %*% x
+
+[[cpp11::register]] doubles solve_y_(const doubles_matrix<> &a,
+                                     const doubles &x) {
+  Mat<double> A = as_Mat(a);
+  Mat<double> X = as_Mat(x);
+
+  return as_doubles(A * X);
+}
+
+// A %*% B %*% A
+
+[[cpp11::register]] doubles_matrix<> sandwich_(const doubles_matrix<> &a,
+                                               const doubles_matrix<> &b) {
+  Mat<double> A = as_Mat(a);
+  Mat<double> B = as_Mat(b);
+
+  Mat<double> res = A * (B * A);
+  return as_doubles_matrix(res);
+}
+
+// eta <- eta.old + rho * eta.upd
+
+[[cpp11::register]] doubles
[…]
