### 2024-07-06 bdfaaba
streamline kendall cor functions
