### 2024-07-16 1daa09a
some refactors to save memory
