### 2024-07-16 5ff84c7
some refactors to save memory
