### 2024-07-16 2629969
initial step halving refactor
