### 2024-07-16 da0a8c8
initial step halving refactor
