+++ src/cpp11.cpp
@@ -131,6 +131,20 @@ extern "C" SEXP _capybara_solve_eta2_(SEXP yadj, SEXP myadj, SEXP offset, SEXP e
     return cpp11::as_sexp(solve_eta2_(cpp11::as_cpp<cpp11::decay_t<const doubles &>>(yadj), cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<> &>>(myadj), cpp11::as_cpp<cpp11::decay_t<const doubles &>>(offset), cpp11::as_cpp<cpp11::decay_t<const doubles &>>(eta)));
   END_CPP11
 }
+// 04_linear_algebra.cpp
+doubles linkinv_(const doubles & eta_r, const std::string & family);
+extern "C" SEXP _capybara_linkinv_(SEXP eta_r, SEXP family) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(linkinv_(cpp11::as_cpp<cpp11::decay_t<const doubles &>>(eta_r), cpp11::as_cpp<cpp11::decay_t<const std::string &>>(family)));
+  END_CPP11
+}
+// 04_linear_algebra.cpp
+double dev_resids_(const doubles & y_r, const doubles & mu_r, const double & theta, const doubles & wt_r, const std::string & family);
+extern "C" SEXP _capybara_dev_resids_(SEXP y_r, SEXP mu_r, SEXP theta, SEXP wt_r, SEXP family) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(dev_resids_(cpp11::as_cpp<cpp11::decay_t<const doubles &>>(y_r), cpp11::as_cpp<cpp11::decay_t<const doubles &>>(mu_r), cpp11::as_cpp<cpp11::decay_t<const double &>>(theta), cpp11::as_cpp<cpp11::decay_t<const doubles &>>(wt_r), cpp11::as_cpp<cpp11::decay_t<const std::string &>>(family)));
+  END_CPP11
+}
 // 05_kendall_correlation.cpp
 double kendall_cor_(const doubles_matrix<> & m);
 extern "C" SEXP _capybara_kendall_cor_(SEXP m) {
@@ -150,6 +164,7 @@ extern "C" {
 static const R_CallMethodDef CallEntries[] = {
     {"_capybara_center_variables_",    (DL_FUNC) &_capybara_center_variables_,    7},
     {"_capybara_crossprod_",           (DL_FUNC) &_capybara_crossprod_,           4},
+    {"_capybara_dev_resids_",          (DL_FUNC) &_capybara_dev_resids_,          5},
     {"_capybara_gamma_",               (DL_FUNC) &_capybara_gamma_,               6},
     {"_capybara_get_alpha_",           (DL_FUNC) &_capybara_get_alpha_,           3},
     {"_capybara_group_sums_",          (DL_FUNC) &_capybara_group_sums_,          3},
@@ -158,6 +173,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_capybara_group_sums_var_",      (DL_FUNC) &_capybara_group_sums_var_,      2},
     {"_capybara_inv_",                 (DL_FUNC) &_capybara_inv_,                 1},
     {"_capybara_kendall_cor_",         (DL_FUNC) &_capybara_kendall_cor_,         1},
+    {"_capybara_linkinv_",             (DL_FUNC) &_capybara_linkinv_,             2},
     {"_capybara_pkendall_",            (DL_FUNC) &_capybara_pkendall_,            2},
     {"_capybara_rank_",                (DL_FUNC) &_capybara_rank_,                1},
     {"_capybara_sandwich_",            (DL_FUNC) &_capybara_sandwich_,            2},
