### 2024-07-16 9719356
continuing step halving refactor
