### 2024-07-16 be0e141
continuing step halving refactor
