### 2024-07-17 b2e92dd
partial refactor to avoid too many R/C++ copies with glm
