+++ src/04_linear_algebra.cpp
@@ -138,48 +138,27 @@ update_beta_eta_(const doubles &old, const doubles &upd, const double &param) {
   return res;
 }
 
-[[cpp11::register]] doubles solve_beta_(const doubles_matrix<> &mx,
-                                        const doubles_matrix<> &mnu,
-                                        const doubles &wtilde,
-                                        const bool &weighted) {
-  Mat<double> X = as_Mat(mx);
-  Mat<double> Y = as_Mat(mnu);
-
-  // Weight the X and Y matrices
-  if (weighted) {
-    Mat<double> w = as_Mat(wtilde);
-    w = sqrt(w);
-    X = X.each_col() % w; // element-wise multiplication
-    Y = Y.each_col() % w;
-  }
-
-  // Solve the system X * beta = Y
-
-  // QR decomposition
+Col<double> solve_beta_(const Mat<double> &MX, const Mat<double> &MNU,
+                                 const Col<double> &w) {
+  Col<double> wtilde = sqrt(w);
 
   Mat<double> Q, R;
 
-  bool computable = qr_econ(Q, R, X);
+  bool computable = qr_econ(Q, R, MX.each_col() % wtilde);
 
   if (!computable) {
     stop("QR decomposition failed");
   } else {
     // backsolve
-    return as_doubles(solve(R, Q.t() * Y));
+    return solve(R, Q.t() * (MNU.each_col() % wtilde));
   }
 }
 
 // eta.upd <- nu - as.vector(Mnu - MX %*% beta.upd)
 
-[[cpp11::register]] doubles solve_eta_(const doubles_matrix<> &mx,
-                                       const doubles_matrix<> &mnu,
-                                       const doubles &nu, const doubles &beta) {
-  Mat<double> MX = as_Mat(mx);
-  Mat<double> MNU = as_Mat(mnu);
-  Mat<double> Nu = as_Mat(nu);
-  Mat<double> Beta = as_Mat(beta);
-
-  return as_doubles(Nu - (MNU - MX * Beta));
+Col<double> solve_eta_(const Mat<double> &MX, const Mat<double> &MNU,
+                        const Col<double> &nu, const Col<double> &beta) {
+  return nu - MNU + MX * beta;
 }
 
 // eta.upd <- yadj - as.vector(Myadj) + offset - eta
@@ -192,153 +171,4 @@ update_beta_eta_(const doubles &old, const doubles &upd, const double &param) {
   Mat<double> Eta = as_Mat(eta);
 
   return as_doubles(Yadj - Myadj + Offset - Eta);
-}
-
-std::string tidy_family(const std::string &family) {
-  // tidy family param
-  std::string fam = family;
-
-  // 1. put all in lowercase
-  std::transform(fam.begin(), fam.end(), fam.begin(),
-                 [](unsigned char c) { return std::tolower(c); });
-
-  // 2. remove numbers
-  fam.erase(std::remove_if(fam.begin(), fam.end(), ::isdigit), fam.end());
-
-  // 3. remove parentheses and everything inside
-  size_t pos = fam.find("(");
-  if (pos != std::string::npos) {
-    fam.erase(pos, fam.size());
-  }
-
-  // 4. replace spaces and dots
-  std::replace(fam.begin(), fam.end(), ' ', '_');
-  std::replace(fam.begin(), fam.end(), '.', '_');
-
-  // 5. trim
-  fam.erase(std::remove_if(fam.begin(), fam.end(), ::isspace), fam.end());
-
-  return fam;
-}
-
-[[cpp11::register]] doubles linkinv_(const doubles &eta_r,
[…]

[…]
