### 2024-07-17 ea9c13d
partial refactor to avoid too many R/C++ copies with glm
