+++ R/cpp11.R
@@ -1,9 +1,5 @@
 # Generated by cpp11: do not edit by hand
 
-center_variables_ <- function(V_r, v_sum_r, w_r, klist, tol, maxiter, sum_v) {
-  .Call(`_capybara_center_variables_`, V_r, v_sum_r, w_r, klist, tol, maxiter, sum_v)
-}
-
 get_alpha_ <- function(p_r, klist, tol) {
   .Call(`_capybara_get_alpha_`, p_r, klist, tol)
 }
@@ -60,32 +56,12 @@ update_nu_ <- function(y, mu, mu_eta) {
   .Call(`_capybara_update_nu_`, y, mu, mu_eta)
 }
 
-solve_beta_ <- function(mx, mnu, wtilde, weighted) {
-  .Call(`_capybara_solve_beta_`, mx, mnu, wtilde, weighted)
-}
-
-solve_eta_ <- function(mx, mnu, nu, beta) {
-  .Call(`_capybara_solve_eta_`, mx, mnu, nu, beta)
-}
-
 solve_eta2_ <- function(yadj, myadj, offset, eta) {
   .Call(`_capybara_solve_eta2_`, yadj, myadj, offset, eta)
 }
 
-linkinv_ <- function(eta_r, family) {
-  .Call(`_capybara_linkinv_`, eta_r, family)
-}
-
-dev_resids_ <- function(y_r, mu_r, theta, wt_r, family) {
-  .Call(`_capybara_dev_resids_`, y_r, mu_r, theta, wt_r, family)
-}
-
-valideta_ <- function(eta_r, family) {
-  .Call(`_capybara_valideta_`, eta_r, family)
-}
-
-validmu_ <- function(mu_r, family) {
-  .Call(`_capybara_validmu_`, mu_r, family)
+feglm_fit_ <- function(beta_r, eta_r, y_r, x_r, nt_r, wt_r, theta, family, control, k_list) {
+  .Call(`_capybara_feglm_fit_`, beta_r, eta_r, y_r, x_r, nt_r, wt_r, theta, family, control, k_list)
 }
 
 kendall_cor_ <- function(m) {
