### 2024-07-17 323b9a9
feglm_fit refactor halfway there
