### 2024-07-17 5c76485
feglm_fit refactor halfway there
