+++ R/internals.R
@@ -1,325 +1,0 @@
-# Transform factor ----
-
-check_factor_ <- function(x) {
-  if (is.factor(x)) {
-    droplevels(x)
-  } else {
-    factor(x)
-  }
-}
-
-# Fitting algorithm (similar to lm.fit) ----
-
-felm_fit_ <- function(y, X, wt, k.list, control) {
-  # Extract control arguments
-  center.tol <- control[["center.tol"]]
-  keep.mx <- control[["keep.mx"]]
-
-  # Generate temporary variables
-  nt <- length(y)
-  MX <- X
-
-  # Centering variables
-  MX <- center_variables_(MX, NA_real_, wt, k.list, center.tol, 10000L, FALSE)
-
-  # Compute the OLS estimate
-  # beta <- as.vector(qr.solve(MX, y, epsilon))
-  beta <- solve_beta_(MX, y, NA_real_, FALSE)
-
-  # Generate result list
-  reslist <- list(
-    coefficients = beta
-  )
-
-  # Update result list
-  if (keep.mx) reslist[["MX"]] <- MX
-
-  # Return result list
-  reslist
-}
-
-# Fitting algorithm (similar to glm.fit) ----
-
-# feglm_fit_ <- function(beta, eta, y, X, wt, k.list, family, control) {
-#   # Extract control arguments
-#   center.tol <- control[["center.tol"]]
-#   dev.tol <- control[["dev.tol"]]
-#   iter.max <- control[["iter.max"]]
-#   keep.mx <- control[["keep.mx"]]
-
-#   # Compute initial quantities for the maximization routine
-#   nt <- length(y)
-#   mu <- family[["linkinv"]](eta)
-#   dev <- sum(family[["dev.resids"]](y, mu, wt))
-#   null.dev <- sum(family[["dev.resids"]](y, mean(y), wt))
-
-#   # Generate temporary variables
-#   Mnu <- as.matrix(numeric(nt))
-#   MX <- X
-
-#   # Start maximization of the log-likelihood
-#   conv <- FALSE
-#   for (iter in seq.int(iter.max)) {
-#     # Store \eta, \beta, and deviance of the previous iteration
-#     eta.old <- eta
-#     beta.old <- beta
-#     dev.old <- dev
-
-#     # Compute weights and dependent variable
-#     mu.eta <- family[["mu.eta"]](eta)
-#     w <- (wt * mu.eta^2) / family[["variance"]](mu)
-#     nu <- (y - mu) / mu.eta
-
-#     # Centering variables
-#     Mnu <- center_variables_(Mnu, nu, w, k.list, center.tol, 10000L, TRUE)
-#     MX <- center_variables_(MX, NA_real_, w, k.list, center.tol, 10000L, FALSE)
-
-#     # Compute update step and update eta
-    
-#     # Step-halving with three checks
-#     # 1. finite deviance
-#     # 2. valid \eta and \mu
-#     # 3. improvement as in glm2
-    
-#     # if (is.integer(y)) y <- as.double(y)
-#     # theta <- ifelse(is.null(family[["theta"]]), 0.0, family[["theta"]])
-#     # sh <- step_halving_(
-#     #   MX, Mnu, nu, w, eta.old, beta.old, y, wt,
-#     #   theta, family[["family"]], dev.old, dev.tol
-#     # )
-
-#     # dev <- sh[["dev"]]
-#     # eta <- sh[["eta"]]
-#     # beta <- sh[["beta"]]
-#     # mu <- sh[["mu"]]
-#     # imp.crit <- sh[["imp.crit"]]
-#     # sh <- NULL
-
[…]
