+++ tests/testthat/test-felm.R
@@ -1,10 +1,27 @@
 test_that("felm works", {
+  load_all()
   m1 <- felm(mpg ~ wt | cyl, mtcars)
   m2 <- lm(mpg ~ wt + as.factor(cyl), mtcars)
 
   expect_equal(round(coef(m1), 5), round(coef(m2)[2], 5))
-  expect_gt(length(fitted(m1)), 0)
-  expect_gt(length(predict(m1)), 0)
-  expect_gt(length(coef(m1)), 0)
-  expect_gt(length(coef(summary(m1))), 0)
+
+  n <- nrow(mtcars)
+  expect_equal(length(fitted(m1)), n)
+  expect_equal(length(predict(m1)), n)
+  expect_equal(length(coef(m1)), 1)
+  expect_equal(length(coef(summary(m1))), 4)
+
+  m1 <- felm(mpg ~ wt + qsec | cyl, mtcars)
+  m2 <- lm(mpg ~ wt + qsec + as.factor(cyl), mtcars)
+
+  expect_equal(round(coef(m1), 5), round(coef(m2)[c(2,3)], 5))
+
+  m1 <- felm(mpg ~ wt + qsec | cyl + am, mtcars)
+  m2 <- lm(mpg ~ wt + qsec + as.factor(cyl) + as.factor(am), mtcars)
+
+  expect_equal(round(coef(m1), 5), round(coef(m2)[c(2, 3)], 5))
+
+  m1 <- felm(mpg ~ wt + qsec | cyl + am | carb, mtcars)
+
+  expect_equal(round(coef(m1), 5), round(coef(m2)[c(2, 3)], 5))
 })
