### 2024-07-20 bf803ab
fix inner ratio for glm iteration stops
