+++ src/05_glm_fit.cpp
@@ -47,9 +47,9 @@ Col<double> link_inv_(const Col<double> &eta, const std::string &fam) {
   if (fam == "gaussian") {
     res = eta;
   } else if (fam == "poisson") {
-    // Col<double> epsilon = 1e-7 * ones<Col<double>>(eta.n_elem);
-    // res = pmax_(exp(eta), epsilon);
     res = exp(eta);
+    // Col<double> epsilon = 2.220446e-16 * ones<Col<double>>(eta.n_elem);
+    // res = pmax_(exp(eta), epsilon);
   } else if (fam == "binomial") {
     // res = exp(eta) / (1.0 + exp(eta));
     res = 1.0 / (1.0 + exp(-eta));
@@ -74,10 +74,10 @@ double dev_resids_(const Col<double> &y, const Col<double> &mu,
   if (fam == "gaussian") {
     res = accu(wt % square(y - mu));
   } else if (fam == "poisson") {
-    uvec p = find(y > 0.0);
+    uvec p = find(y > 0);
     Col<double> r = mu % wt;
     r(p) = wt(p) % (y(p) % log(y(p) / mu(p)) - (y(p) - mu(p)));
-    res = 2.0 * accu(r);
+    res = 2 * accu(r);
   } else if (fam == "binomial") {
     uvec p = find(y != 0.0);
     uvec q = find(y != 1.0);
@@ -134,7 +134,7 @@ bool valid_mu_(const Col<double> &mu, const std::string &fam) {
   if (fam == "gaussian") {
     res = true;
   } else if (fam == "poisson") {
-    res = is_finite(mu) && all(mu > 0.0);
+    res = is_finite(mu) && all(mu > 0);
   } else if (fam == "binomial") {
     res = is_finite(mu) && all(mu > 0.0 && mu < 1.0);
   } else if (fam == "gamma") {
@@ -159,6 +159,8 @@ Col<double> mu_eta_(Col<double> &eta, const std::string &fam) {
     res.ones();
   } else if (fam == "poisson") {
     res = exp(eta);
+    // Col<double> epsilon = 2.220446e-16 * ones<Col<double>>(eta.n_elem);
+    // res = pmax_(exp(eta), epsilon);
   } else if (fam == "binomial") {
     res = 1.0 / (2.0 + exp(eta) + exp(-eta));
   } else if (fam == "gamma") {
@@ -174,8 +176,8 @@ Col<double> mu_eta_(Col<double> &eta, const std::string &fam) {
   return res;
 }
 
-Col<double> variance_(const Col<double> &mu,
-                      const double &theta, const std::string &fam) {
+Col<double> variance_(const Col<double> &mu, const double &theta,
+                      const std::string &fam) {
   Col<double> res(mu.n_elem);
 
   if (fam == "gaussian") {
@@ -197,19 +199,20 @@ Col<double> variance_(const Col<double> &mu,
   return res;
 }
 
-[[cpp11::register]] list feglm_fit_(
-    const doubles &beta_r, const doubles &eta_r, const doubles &y_r,
-    const doubles_matrix<> &x_r, const double &nt, const doubles &wt_r,
-    const double &theta, const std::string &family, const list &control,
-    const list &k_list) {
+[[cpp11::register]] list feglm_fit_(const doubles &beta_r, const doubles &eta_r,
+                                    const doubles &y_r,
+                                    const doubles_matrix<> &x_r,
+                                    const doubles &wt_r,
+                                    const double &theta,
+                                    const std::string &family,
[…]
