### 2024-07-20 707b7cf
fix inner ratio for glm iteration stops
