### 2024-07-20 7deae73
correct fixed_effects() with C++ changes
