### 2024-07-20 4d41597
correct fixed_effects() with C++ changes
