### 2024-07-20 7fe44d0
remove functions with large overhead
