+++ src/cpp11.cpp
@@ -40,69 +40,6 @@ extern "C" SEXP _capybara_group_sums_cov_(SEXP M_r, SEXP N_r, SEXP jlist) {
     return cpp11::as_sexp(group_sums_cov_(cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<> &>>(M_r), cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<> &>>(N_r), cpp11::as_cpp<cpp11::decay_t<const list &>>(jlist)));
   END_CPP11
 }
-// 04_linear_algebra.cpp
-doubles_matrix<> gamma_(const doubles_matrix<> & mx, const doubles_matrix<> & hessian, const doubles_matrix<> & j, const doubles_matrix<> & ppsi, const doubles & v, const SEXP & nt_full);
-extern "C" SEXP _capybara_gamma_(SEXP mx, SEXP hessian, SEXP j, SEXP ppsi, SEXP v, SEXP nt_full) {
-  BEGIN_CPP11
-    return cpp11::as_sexp(gamma_(cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<> &>>(mx), cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<> &>>(hessian), cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<> &>>(j), cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<> &>>(ppsi), cpp11::as_cpp<cpp11::decay_t<const doubles &>>(v), cpp11::as_cpp<cpp11::decay_t<const SEXP &>>(nt_full)));
-  END_CPP11
-}
-// 04_linear_algebra.cpp
-doubles_matrix<> inv_(const doubles_matrix<> & h);
-extern "C" SEXP _capybara_inv_(SEXP h) {
-  BEGIN_CPP11
-    return cpp11::as_sexp(inv_(cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<> &>>(h)));
-  END_CPP11
-}
-// 04_linear_algebra.cpp
-int rank_(const doubles_matrix<> & x);
-extern "C" SEXP _capybara_rank_(SEXP x) {
-  BEGIN_CPP11
-    return cpp11::as_sexp(rank_(cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<> &>>(x)));
-  END_CPP11
-}
-// 04_linear_algebra.cpp
-doubles solve_bias_(const doubles & beta_uncorr, const doubles_matrix<> & hessian, const double & nt, const doubles & b);
-extern "C" SEXP _capybara_solve_bias_(SEXP beta_uncorr, SEXP hessian, SEXP nt, SEXP b) {
-  BEGIN_CPP11
-    return cpp11::as_sexp(solve_bias_(cpp11::as_cpp<cpp11::decay_t<const doubles &>>(beta_uncorr), cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<> &>>(hessian), cpp11::as_cpp<cpp11::decay_t<const double &>>(nt), cpp11::as_cpp<cpp11::decay_t<const doubles &>>(b)));
-  END_CPP11
-}
-// 04_linear_algebra.cpp
-doubles solve_y_(const doubles_matrix<> & a, const doubles & x);
-extern "C" SEXP _capybara_solve_y_(SEXP a, SEXP x) {
-  BEGIN_CPP11
-    return cpp11::as_sexp(solve_y_(cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<> &>>(a), cpp11::as_cpp<cpp11::decay_t<const doubles &>>(x)));
-  END_CPP11
-}
[…]
