### 2024-07-20 2849790
remove functions with large overhead
