### 2024-07-21 0d31f8c
mimic logit from base R
