+++ src/05_glm_fit.cpp
@@ -41,18 +41,80 @@ std::string tidy_family_(const std::string &family) {
 //   return res;
 // }
 
+// static const double THRESH = 30.;
+// static const double MTHRESH = -30.;
+// static const double DBL_EPSILON = std::numeric_limits<double>::epsilon();
+// static const double INVEPS = 1 / DBL_EPSILON;
+
+// Col<double> link_inv_logit_(const Col<double> &x) {
+//   Col<double> res(x.n_elem);
+
+//   uword i, n = x.n_elem;
+//   for (i = 0; i < n; ++i) {
+//     if (x(i) < MTHRESH) {
+//       res(i) = DBL_EPSILON;
+//     } else if (x(i) > THRESH) {
+//       res(i) = INVEPS;
+//     } else {
+//       double y = exp(x(i));
+//       res(i) = y / (1 + y);
+//     }
+//   }
+
+//   return res;
+// }
+
+Col<double> link_inv_logit_(const Col<double> &x) {
+  Col<double> y = exp(x);
+  return y / (1 + y);
+}
+
+// Adapted from binomial_dev_resids()
+// in R base it can be found in src/library/stats/src/family.c
+// unfortunately the functions that work with a SEXP won't work with a Col<>
+Col<double> dev_resids_logit_(const Col<double> &y, const Col<double> &mu) {
+  Col<double> res(y.n_elem, fill::zeros);
+
+  uvec p = find(y != 0);
+  res(p) = y(p) % log(y(p) / mu(p));
+
+  return res;
+}
+
+// Col<double> mu_eta_logit_(const Col<double> &x) {
+//   Col<double> res(x.n_elem);
+
+//   uword i, n = x.n_elem;
+
+//   for (i = 0; i < n; ++i) {
+//     double opexp = 1 + exp(x(i));
+    
+//     if ((x(i) > THRESH) || (x(i) < MTHRESH)){
+//       res(i) = DBL_EPSILON;
+//     } else {
+//       res(i) = exp(x(i)) / (opexp * opexp);
+//     }
+//   }
+
+//   return res;
+// }
+
+Col<double> mu_eta_logit_(const Col<double> &x) {
+  Col<double> y = exp(x);
+  return y / square(1 + y);
+}
+
 Col<double> link_inv_(const Col<double> &eta, const std::string &fam) {
   Col<double> res(eta.n_elem);
   
   if (fam == "gaussian") {
     res = eta;
   } else if (fam == "poisson") {
-    res = exp(eta);
     // Col<double> epsilon = 2.220446e-16 * ones<Col<double>>(eta.n_elem);
     // res = pmax_(exp(eta), epsilon);
+    res = exp(eta);
   } else if (fam == "binomial") {
-    // res = exp(eta) / (1.0 + exp(eta));
-    res = 1.0 / (1.0 + exp(-eta));
+    res = link_inv_logit_(eta);
   } else if (fam == "gamma") {
     res = 1.0 / eta;
   } else if (fam == "inverse_gaussian") {
@@ -79,13 +141,8 @@ double dev_resids_(const Col<double> &y, const Col<double> &mu,
     r(p) = wt(p) % (y(p) % log(y(p) / mu(p)) - (y(p) - mu(p)));
     res = 2 * accu(r);
   } else if (fam == "binomial") {
-    uvec p = find(y != 0.0);
-    uvec q = find(y != 1.0);
-    Col<double> r = y / mu;
-    Col<double> s = (1.0 - y) / (1.0 - mu);
-    r(p) = log(r(p));
-    s(q) = log(s(q));
-    res = 2.0 * accu(wt % (y % r + (1.0 - y) % s));
+    res = 2 * accu(wt % (dev_resids_logit_(y, mu) +
[…]
