### 2024-07-21 32a31fc
simplify notation for step halving
