### 2024-07-21 d873ec6
fix logit convergence issue
