+++ src/05_glm_fit.cpp
@@ -27,46 +27,34 @@ std::string tidy_family_(const std::string &family) {
   return fam;
 }
 
-// Pairwise-maximum function
-// Col<double> pmax_(const Col<double> &x, const Col<double> &y) {
-//   Col<double> res(x.n_elem);
+Col<double> link_inv_gaussian_(const Col<double> &x) {
+  return x;
+}
 
-//   // for (int i = 0; i < x.n_elem; ++i) {
-//   //   res(i) = std::max(x(i), y(i));
-//   // }
-
-//   std::transform(x.begin(), x.end(), y.begin(), res.begin(),
-//                  [](double a, double b) { return std::max(a, b); });
-
-//   return res;
-// }
-
-// static const double THRESH = 30.;
-// static const double MTHRESH = -30.;
-// static const double DBL_EPSILON = std::numeric_limits<double>::epsilon();
-// static const double INVEPS = 1 / DBL_EPSILON;
-
-// Col<double> link_inv_logit_(const Col<double> &x) {
-//   Col<double> res(x.n_elem);
-
-//   uword i, n = x.n_elem;
-//   for (i = 0; i < n; ++i) {
-//     if (x(i) < MTHRESH) {
-//       res(i) = DBL_EPSILON;
-//     } else if (x(i) > THRESH) {
-//       res(i) = INVEPS;
-//     } else {
-//       double y = exp(x(i));
-//       res(i) = y / (1 + y);
-//     }
-//   }
-
-//   return res;
-// }
+Col<double> link_inv_poisson_(const Col<double> &x) {
+  return exp(x);
+}
 
 Col<double> link_inv_logit_(const Col<double> &x) {
   Col<double> y = exp(x);
   return y / (1 + y);
+}
+
+Col<double> link_inv_gamma_(const Col<double> &x) {
+  return 1 / x;
+}
+
+Col<double> link_inv_invgaussian_(const Col<double> &x) {
+  return 1 / sqrt(x);
+}
+
+Col<double> link_inv_negbin_(const Col<double> &x) {
+  return exp(x);
+}
+
+double dev_resids_gaussian_(const Col<double> &y, const Col<double> &mu,
+                                const Col<double> &wt) {
+  return accu(wt % square(y - mu));
 }
 
 double dev_resids_poisson_(const Col<double> &y, const Col<double> &mu,
@@ -121,46 +109,46 @@ double dev_resids_negbin_(const Col<double> &y, const Col<double> &mu,
   return 2 * accu(r);
 }
 
-// Col<double> mu_eta_logit_(const Col<double> &x) {
-//   Col<double> res(x.n_elem);
+Col<double> mu_eta_gaussian_(const Col<double> &x) {
+  return ones<Col<double>>(x.n_elem);
+}
 
-//   uword i, n = x.n_elem;
-
-//   for (i = 0; i < n; ++i) {
-//     double opexp = 1 + exp(x(i));
-    
-//     if ((x(i) > THRESH) || (x(i) < MTHRESH)){
-//       res(i) = DBL_EPSILON;
-//     } else {
-//       res(i) = exp(x(i)) / (opexp * opexp);
-//     }
-//   }
-
-//   return res;
-// }
+Col<double> mu_eta_poisson_(const Col<double> &x) {
+  return exp(x);
+}
 
 Col<double> mu_eta_logit_(const Col<double> &x) {
   Col<double> y = exp(x);
   return y / square(1 + y);
 }
 
+Col<double> mu_eta_gamma_(const Col<double> &x) {
+  return -1 / square(x);
+}
+
+Col<double> mu_eta_invgaussian_(const Col<double> &x) {
+  return -1 / (2 * pow(x, 1.5));
+}
+
[…]
