### 2024-07-21 f1ad80c
streamline code
