### 2024-07-21 e5e04ad
streamline code
