### 2024-07-22 6cf87e1
v0.6.0 avoids multiple R <-> C++ data copies
