+++ R/cpp11.R
@@ -1,5 +1,9 @@
 # Generated by cpp11: do not edit by hand
 
+center_variables_r_ <- function(V_r, w_r, klist, tol, maxiter) {
+  .Call(`_capybara_center_variables_r_`, V_r, w_r, klist, tol, maxiter)
+}
+
 get_alpha_ <- function(p_r, klist, tol) {
   .Call(`_capybara_get_alpha_`, p_r, klist, tol)
 }
