### 2024-07-22 1cb54e7
v0.6.0 avoids multiple R <-> C++ data copies
