+++ src/01_center_variables.cpp
@@ -68,3 +68,11 @@ Mat<double> center_variables_(const Mat<double> &V, const Col<double> &w,
   // Return matrix with centered variables
   return C;
 }
+
+[[cpp11::register]] doubles_matrix<> center_variables_r_(
+    const doubles_matrix<> &V_r, const doubles &w_r,
+    const list &klist, const double &tol, const int &maxiter) {
+  return as_doubles_matrix(
+    center_variables_(as_Mat(V_r), as_Mat(w_r), klist, tol, maxiter)
+  );
+}
