+++ NEWS.md
@@ -1,3 +1,11 @@
+# capybara 0.6.0
+
+* Moves all the heavy computation to C++ using Armadillo and it exports the 
+  results to R. Previously, there were multiple data copies between R and C++
+  that added overhead to the computations.
+* For a future release, I may rewrite the offset and APES computation to C++,
+  but with those the overhead is minimal.
+
 # capybara 0.5.2
 
 * Uses an O(n log(n)) algorithm to compute the Kendall correlation for the
