+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: capybara
 Type: Package
 Title: Fast and Memory Efficient Fitting of Linear Models With High-Dimensional
     Fixed Effects
-Version: 0.5.2
+Version: 0.6.0
 Authors@R: c(
     person(
         given = "Mauricio",

[…]
