+++ src/00_main.h
@@ -27,3 +27,20 @@ Col<double> solve_eta_(const Mat<double> &MX, const Mat<double> &MNU,
 Mat<double> crossprod_(const Mat<double> &X, const Col<double> &w, const int &n,
                        const int &p, const bool &weighted,
                        const bool &root_weights);
+
+std::string tidy_family_(const std::string &family);
+
+Col<double> link_inv_(const Col<double> &eta, const std::string &fam);
+
+double dev_resids_(const Col<double> &y, const Col<double> &mu,
+                   const double &theta, const Col<double> &wt,
+                   const std::string &fam);
+
+Col<double> mu_eta_(Col<double> &eta, const std::string &fam);
+
+Col<double> variance_(const Col<double> &mu, const double &theta,
+                      const std::string &fam);
+
+bool valid_eta_(const Col<double> &eta, const std::string &fam);
+
+bool valid_mu_(const Col<double> &mu, const std::string &fam);
