+++ R/bias_corr.R
@@ -16,7 +16,7 @@
 #'  weakly exogenous regressors, e.g. lagged outcome variables, we suggest to
 #'  choose a bandwidth between one and four. Note that the order of factors to
 #'  be partialed out is important for bandwidths larger than zero.
-#' @param panel.structure a string equal to \code{"classic"} or \code{"network"}
+#' @param panel_structure a string equal to \code{"classic"} or \code{"network"}
 #'  which determines the structure of the panel used. \code{"classic"} denotes
 #'  panel structures where for example the same cross-sectional units are
 #'  observed several times (this includes pseudo panels). \code{"network"}
@@ -59,7 +59,7 @@
 bias_corr <- function(
     object = NULL,
     L = 0L,
-    panel.structure = c("classic", "network")) {
+    panel_structure = c("classic", "network")) {
   # Check validity of 'object'
   if (is.null(object)) {
     stop("'object' has to be specified.", call. = FALSE)
@@ -67,20 +67,20 @@ bias_corr <- function(
     stop("'bias_corr' called on a non-'feglm' object.", call. = FALSE)
   }
 
-  # Check validity of 'panel.structure'
-  panel.structure <- match.arg(panel.structure)
+  # Check validity of 'panel_structure'
+  panel_structure <- match.arg(panel_structure)
 
   # Extract model information
-  beta.uncorr <- object[["coefficients"]]
+  beta_uncorr <- object[["coefficients"]]
   control <- object[["control"]]
   data <- object[["data"]]
   eps <- .Machine[["double.eps"]]
   family <- object[["family"]]
   formula <- object[["formula"]]
   lvls_k <- object[["lvls_k"]]
-  nms.sp <- names(beta.uncorr)
+  nms.sp <- names(beta_uncorr)
   nt <- object[["nobs"]][["nobs"]]
-  k.vars <- names(lvls_k)
+  k_vars <- names(lvls_k)
   k <- length(lvls_k)
 
   # Check if binary choice model
@@ -100,11 +100,11 @@ bias_corr <- function(
   }
 
   # Check if provided object matches requested panel structure
-  if (panel.structure == "classic") {
+  if (panel_structure == "classic") {
     if (!(k %in% c(1L, 2L))) {
       stop(
         paste(
-          "panel.structure == 'classic' expects a one- or two-way fixed",
+          "panel_structure == 'classic' expects a one- or two-way fixed",
           "effect model."
         ),
         call. = FALSE
@@ -114,7 +114,7 @@ bias_corr <- function(
     if (!(k %in% c(2L, 3L))) {
       stop(
         paste(
-          "panel.structure == 'network' expects a two- or three-way fixed",
+          "panel_structure == 'network' expects a two- or three-way fixed",
           "effects model."
         ),
         call. = FALSE
@@ -129,17 +129,17 @@ bias_corr <- function(
   wt <- object[["weights"]]
 
   # Generate auxiliary list of indexes for different sub panels
-  k.list <- get_index_list_(k.vars, data)
+  k_list <- get_index_list_(k_vars, data)
 
   # Compute derivatives and weights
   eta <- object[["eta"]]
   mu <- family[["linkinv"]](eta)
-  mu.eta <- family[["mu.eta"]](eta)
+  mu_eta <- family[["mu.eta"]](eta)
[…]
