### 2024-07-23 e5e0b53
correct adj rsq
