### 2024-07-23 1f3d709
correct adj rsq
