+++ dev/test-get_index_list.R
@@ -1,0 +1,18 @@
+test_that("get_index_list works", {
+  get_index_list_r_ <- function(k.vars, data) {
+    indexes <- seq.int(0L, nrow(data) - 1L)
+    lapply(k.vars, function(x, indexes, data) {
+      split(indexes, data[[x]])
+    }, indexes = indexes, data = data)
+  }
+
+  expect_equal(
+    get_index_list_(names(mtcars), mtcars),
+    get_index_list_r_(names(mtcars), mtcars)
+  )
+
+  expect_equal(
+    get_index_list_(names(iris), iris),
+    get_index_list_r_(names(iris), iris)
+  )
+})
