+++ dev/04_linear_algebra.cpp
@@ -1,369 +1,0 @@
-#include "00_main.h"
-
-// helper to convert std::vector<double> to doubles_matrix<>
-writable::doubles_matrix<> stdvec_to_dblmatrix_(const std::vector<double> &vec,
-                                                int rows, int cols) {
-  writable::doubles_matrix<> mat(rows, cols);
-
-#pragma omp parallel for
-  for (int i = 0; i < rows; ++i) {
-    for (int j = 0; j < cols; ++j) {
-      mat(i, j) = vec[i * cols + j];
-    }
-  }
-  return mat;
-}
-
-// Y <- crossprod(X)
-// Y <- t(X) %*% X
-
-[[cpp11::register]] doubles_matrix<> crossprod_(const doubles_matrix<> &x,
-                                                const doubles &w,
-                                                const bool &weighted,
-                                                const bool &root_weights) {
-  int N = x.nrow();
-  int P = x.ncol();
-
-  const double *x_data = REAL(x.data());
-  const double *w_data = REAL(w);
-
-  std::vector<double> res_vec(P * P, 0.0);
-
-  if (!weighted) {
-    // Multiply X.t() by X
-#pragma omp parallel for
-    for (int n = 0; n < N; n++) {
-      for (int p = 0; p < P; p++) {
-        for (int q = 0; q < P; q++) {
-          double temp = x_data[n + N * p] * x_data[n + N * q];
-#pragma omp atomic
-          res_vec[p * P + q] += temp;
-        }
-      }
-    }
-  } else {
-    std::vector<double> weights(N);
-
-    if (!root_weights) {
-#pragma omp parallel for
-      for (int n = 0; n < N; n++) {
-        weights[n] = w_data[n];
-      }
-    } else {
-#pragma omp parallel for
-      for (int n = 0; n < N; n++) {
-        weights[n] = sqrt(w_data[n]);
-      }
-    }
-
-    // Multiply weighted X.t() by weighted X
-#pragma omp parallel for
-    for (int n = 0; n < N; n++) {
-      for (int p = 0; p < P; p++) {
-        for (int q = 0; q < P; q++) {
-          double temp =
-              x_data[n + N * p] * x_data[n + N * q] * pow(weights[n], 2);
-#pragma omp atomic
-          res_vec[p * P + q] += temp;
-        }
-      }
-    }
-  }
-
-  // writable::doubles_matrix<> res(P, P);
-
-  // #pragma omp parallel for
-  // for (int p = 0; p < P; p++) {
-  //   for (int q = 0; q < P; q++) {
-  //     res(p, q) = res_vec[p * P + q];
-  //   }
-  // }
-
-  // return res;
-
-  return stdvec_to_dblmatrix_(res_vec, P, P);
-}
-
-// WinvJ < -solve(object[["Hessian"]] / nt.full, J)
-// Gamma < -(MX %*% WinvJ - PPsi) * v / nt.full
-// V < -crossprod(Gamma)
-
-[[cpp11::register]] doubles_matrix<>
-gamma_(const doubles_matrix<> &mx, const doubles_matrix<> &hessian,
-       const doubles_matrix<> &j, const doubles_matrix<> &ppsi,
-       const doubles &v, const SEXP &nt_full) {
-  // Types conversion
-  Mat<double> MX = as_Mat(mx);
-  Mat<double> H = as_Mat(hessian);
-  Mat<double> J = as_Mat(j);
-  Mat<double> PPsi = as_Mat(ppsi);
-  Mat<double> V = as_Mat(v);
-
-  double inv_N = 1.0 / as_cpp<double>(nt_full);
-
[…]
