### 2024-07-23 886ecb1
data.table reduces ram usage but modifies the original data
