### 2024-07-25 #645 @pachadotdev
hi @jooolia
I think the software is ready for review as is
