+++ README.Rmd
@@ -26,35 +26,31 @@ knitr::opts_chunk$set(
 
 tldr; If you have a 2-4GB dataset and you need to estimate a (generalized)
 linear model with a large number of fixed effects, this package is for you. It
-works with larger datasets as well.
+works with larger datasets as well and facilites computing clustered standard
+errors.
 
-Capybara is a fast and small footprint software that provides efficient
-functions for demeaning variables before conducting a GLM estimation via
-Iteratively Weighted Least Squares (IWLS). This technique is particularly useful
-when estimating linear models with multiple group fixed effects. It is a fork
-of the excellent Alpaca package created and maintained by
-[Dr. Amrei Stammann](https://github.com/amrei-stammann).
-
-The software can estimate GLMs from the Exponential Family and also Negative
-Binomial models but the focus will be the Poisson estimator because it is the
-one used for structural counterfactual analysis in International Trade. It is
-relevant to add that the IWLS estimator is equivalent with the PPML estimator
-from Santos-Silva et al. 2006
+'capybara' is a fast and small footprint software that provides efficient
+functions for demeaning variables before conducting a GLM estimation. This
+technique is particularly useful when estimating linear models with multiple
+group fixed effects. It is a fork of the excellent Alpaca package created and
+maintained by [Dr. Amrei Stammann](https://github.com/amrei-stammann). The
+software can estimate Exponential Family models (e.g., Poisson) and Negative
+Binomial models.
 
 Traditional QR estimation can be unfeasible due to additional memory
-requirements. The method, which is based on Halperin 1962 article on vector
-projections offers important time and memory savings without compromising
-numerical stability in the estimation process.
+requirements. The method, which is based on Halperin (1962) vector projections
+offers important time and memory savings without compromising numerical
+stability in the estimation process.
 
-The software heavily borrows from Gaure 20213 and Stammann 2018 works on the
-OLS and IWLS estimator with large k-way fixed effects (i.e., the lfe and alpaca
-packages). The differences are that Capybara does not use C nor Rcpp code,
-instead it uses the cpp11 and
-[cpp11armadillo](https://github.com/pachadotdev/cpp11armadillo) packages.
+The software heavily borrows from Gaure (2013) and Stammann (2018) works on
+OLS and GLM estimation with large fixed effects implemented in the 'lfe' and
+'alpaca' packages. The differences are that 'capybara' does not use C nor Rcpp
+code, instead it uses cpp11 and
+[cpp11armadillo](https://github.com/pachadotdev/cpp11armadillo).
 
-The summary tables are nothing like R's default and borrow from Stata outputs.
-I have also provided integrations with broom to facilitate the inclusion of
-statistical tables in Quarto/Jupyter notebooks.
+The summary tables borrow from Stata outputs. I have also provided integrations
+with 'broom' to facilitate the inclusion of statistical tables in Quarto/Jupyter
+notebooks.
 
 If this software is useful to you, please consider donating on
 [Buy Me A Coffee](https://buymeacoffee.com/pacha). All donations will
@@ -74,47 +70,39 @@ See the documentation in progress: https://pacha.dev/capybara.
 
 ## Design choices
 
-Capybara uses C++ and vectorized R operations to address bottlenecks. Some
-parts of the code use dplyr, which allows me to write code that is easier to
-understand and it works very well to performed grouped operations.
-
-I tried to implement an important lesson between v0.1 and v0.2: "He who gives up
-[code] safety for [code] speed deserves neither." (Wickham, 2014).
+Capybara uses C++ and vectorized R operations to address bottlenecks where
+possible. Some parts of the code use 'dplyr', which allows me to write code that
+is easier to understand and it works well to performed grouped operations. The
[…]
