### 2024-07-30 #645 @pachadotdev
> Dear @pachadotdev, thanks for updating to address the statistical aspects of the package. Currently there are some srrstatsTODO tags in your srr-stats-standards.R file and we note that "software to be submitted for review must contain no @srrstatsTODO tags" (https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-srr).
>
> Could you also edit your first comment in this issue so that it contains the comment here [#645 (comment)](https://github.com/ropensci/software-review/issues/645#issuecomment-2209395422) (e.g. replace the content with the statistical issue content? )
>
> Thanks, Julia

thanks!

I updated the comment. I will need 3-4 weeks to finish all the srrstatsTODO
