### 2024-08-02 #645 @adamhsparks
Hi, @pachadotdev, I'm taking over where @jooolia is leaving off as EIC. Please let me know what assistance you may need from me.
