### 2024-08-06 8fda07c
partial changes for ropensci standard
