+++ R/helpers.R
@@ -1,5 +1,7 @@
-# Checks if variable is a factor and transforms if necessary ---
-
+#' @title Transform factor
+#' @description Checks if variable is a factor and transforms if necessary
+#' @param x Variable to be checked
+#' @noRd
 check_factor_ <- function(x) {
   if (is.factor(x)) {
     droplevels(x)
@@ -8,67 +10,87 @@ check_factor_ <- function(x) {
   }
 }
 
-# Higher-order partial derivatives ----
-
-second_order_derivative_ <- function(eta, f, family) {
+#' @title Second order derivative
+#' @description Helper for the partial_mu_eta function
+#' @param eta Eta value
+#' @param mu_eta Mu.eta value
+#' @param family Family object
+#' @noRd
+second_order_derivative_ <- function(eta, mu_eta, family) {
   link <- family[["link"]]
   linkinv_eta <- family[["linkinv"]](eta)
 
   if (link == "logit") {
-    return(f * (1.0 - 2.0 * linkinv_eta))
+    return(mu_eta * (1.0 - 2.0 * linkinv_eta))
   } else if (link == "probit") {
-    return(-eta * f)
+    return(-eta * mu_eta)
   } else if (link == "cloglog") {
-    return(f * (1.0 - exp(eta)))
+    return(mu_eta * (1.0 - exp(eta)))
   } else {
-    return(-2.0 * eta / (1.0 + eta^2) * f)
+    return(-2.0 * eta / (1.0 + eta^2) * mu_eta)
   }
 }
 
-third_order_derivative_ <- function(eta, f, family) {
+#' @title Third order derivative
+#' @description Helper for the partial_mu_eta function
+#' @param eta Eta value
+#' @param mu_eta Mu.eta value
+#' @param family Family object
+#' @noRd
+third_order_derivative_ <- function(eta, mu_eta, family) {
   link <- family[["link"]]
   linkinv_eta <- family[["linkinv"]](eta)
 
   if (link == "logit") {
-    return(f * ((1.0 - 2.0 * linkinv_eta)^2 - 2.0 * f))
+    return(mu_eta * ((1.0 - 2.0 * linkinv_eta)^2 - 2.0 * mu_eta))
   } else if (link == "probit") {
-    return((eta^2 - 1.0) * f)
+    return((eta^2 - 1.0) * mu_eta)
   } else if (link == "cloglog") {
-    return(f * (1.0 - exp(eta)) * (2.0 - exp(eta)) - f)
+    return(mu_eta * (1.0 - exp(eta)) * (2.0 - exp(eta)) - mu_eta)
   } else {
-    return((6.0 * eta^2 - 2.0) / (1.0 + eta^2)^2 * f)
+    return((6.0 * eta^2 - 2.0) / (1.0 + eta^2)^2 * mu_eta)
   }
 }
 
+#' @title Second or third order derivative
+#' @description Computes the second or third order derivative of the link function
+#' @param eta Linear predictor
+#' @param family Family object
+#' @param order Order of the derivative (2 or 3)
+#' @noRd
 partial_mu_eta_ <- function(eta, family, order) {
   # Safeguard eta if necessary
   if (family[["link"]] != "logit") {
     eta <- family[["linkfun"]](family[["linkinv"]](eta))
   }
 
-  f <- family[["mu.eta"]](eta)
+  mu_eta <- family[["mu.eta"]](eta)
 
   if (order == 2L) {
-    return(second_order_derivative_(eta, f, family))
+    return(second_order_derivative_(eta, mu_eta, family))
   } else {
-    return(third_order_derivative_(eta, f, family))
+    return(third_order_derivative_(eta, mu_eta, family))
   }
 }
 
[…]
