+++ DESCRIPTION
@@ -37,10 +37,10 @@ LazyData: true
 RoxygenNote: 7.3.1
 Encoding: UTF-8
 NeedsCompilation: yes
-LinkingTo: cpp11, cpp11armadillo
+LinkingTo: cpp11, armadillo
 VignetteBuilder: knitr
 Config/testthat/edition: 3
 Remotes: 
-    pachadotdev/cpp11armadillo,
+    pachadotdev/armadillo,
     ropenscilabs/srr
 Roxygen: list(markdown = TRUE, roclets = c("namespace", "rd", "srr::srr_stats_roclet"))

[…]
