### 2024-08-06 3797a4d
partial changes for ropensci standard
