+++ NAMESPACE
@@ -32,6 +32,7 @@ export(apes)
 export(augment)
 export(bias_corr)
 export(feglm)
+export(feglm_control)
 export(felm)
 export(fenegbin)
 export(fepoisson)
