+++ src/05_glm_fit.cpp
@@ -27,38 +27,30 @@ std::string tidy_family_(const std::string &family) {
   return fam;
 }
 
-Col<double> link_inv_gaussian_(const Col<double> &eta) {
-  return eta;
-}
+Col<double> link_inv_gaussian_(const Col<double> &eta) { return eta; }
 
-Col<double> link_inv_poisson_(const Col<double> &eta) {
-  return exp(eta);
-}
+Col<double> link_inv_poisson_(const Col<double> &eta) { return exp(eta); }
 
 Col<double> link_inv_logit_(const Col<double> &eta) {
   Col<double> expeta = exp(eta);
   return expeta / (1 + expeta);
 }
 
-Col<double> link_inv_gamma_(const Col<double> &eta) {
-  return 1 / eta;
-}
+Col<double> link_inv_gamma_(const Col<double> &eta) { return 1 / eta; }
 
 Col<double> link_inv_invgaussian_(const Col<double> &eta) {
   return 1 / sqrt(eta);
 }
 
-Col<double> link_inv_negbin_(const Col<double> &eta) {
-  return exp(eta);
-}
+Col<double> link_inv_negbin_(const Col<double> &eta) { return exp(eta); }
 
 double dev_resids_gaussian_(const Col<double> &y, const Col<double> &mu,
-                                const Col<double> &wt) {
+                            const Col<double> &wt) {
   return accu(wt % square(y - mu));
 }
 
 double dev_resids_poisson_(const Col<double> &y, const Col<double> &mu,
-                                const Col<double> &wt) {
+                           const Col<double> &wt) {
   Col<double> r = mu % wt;
 
   uvec p = find(y > 0);
@@ -70,7 +62,8 @@ double dev_resids_poisson_(const Col<double> &y, const Col<double> &mu,
 // Adapted from binomial_dev_resids()
 // in R base it can be found in src/library/stats/src/family.c
 // unfortunately the functions that work with a SEXP won't work with a Col<>
-double dev_resids_logit_(const Col<double> &y, const Col<double> &mu, const Col<double> &wt) {
+double dev_resids_logit_(const Col<double> &y, const Col<double> &mu,
+                         const Col<double> &wt) {
   Col<double> r(y.n_elem, fill::zeros);
   Col<double> s(y.n_elem, fill::zeros);
 
@@ -83,9 +76,9 @@ double dev_resids_logit_(const Col<double> &y, const Col<double> &mu, const Col<
 }
 
 double dev_resids_gamma_(const Col<double> &y, const Col<double> &mu,
-                              const Col<double> &wt) {
+                         const Col<double> &wt) {
   Col<double> r = y / mu;
-  
+
   uvec p = find(y == 0);
   r.elem(p).fill(1.0);
   r = wt % (log(r) - (y - mu) / mu);
@@ -94,14 +87,14 @@ double dev_resids_gamma_(const Col<double> &y, const Col<double> &mu,
 }
 
 double dev_resids_invgaussian_(const Col<double> &y, const Col<double> &mu,
-                                   const Col<double> &wt) {
+                               const Col<double> &wt) {
   return accu(wt % square(y - mu) / (y % square(mu)));
 }
 
 double dev_resids_negbin_(const Col<double> &y, const Col<double> &mu,
-                               const double &theta, const Col<double> &wt) {
+                          const double &theta, const Col<double> &wt) {
   Col<double> r = y;
-  
+
   uvec p = find(y < 1);
   r.elem(p).fill(1.0);
   r = wt % (y % log(r / mu) - (y + theta) % log((y + theta) / (mu + theta)));
@@ -113,42 +106,30 @@ Col<double> mu_eta_gaussian_(const Col<double> &eta) {
[…]
