### 2024-08-10 e32645c
add cores documentation
