+++ tests/testthat/test-fepoisson.R
@@ -48,33 +48,33 @@ test_that("fepoisson is similar to fixest", {
   expect_output(summary_fisher_(smod))
 
   # unique(trade_panel$year)
-  trade_panel_2006 <- trade_panel[trade_panel$year == 2006, ]
+  trade_panel_2 <- trade_panel[trade_panel$year %in% c(2002,2006), ]
 
   if (identical(Sys.info()[['user']], 'pacha')) {
     t_fepoisson <- rep(0,10)
-    for (i in 1:10) {
+    # for (i in 1:10) {
       t1 <- Sys.time()
       fit <- fepoisson(
         trade ~ log_dist + lang + cntg + clny | exp_year + imp_year,
-        trade_panel_2006
+        trade_panel_2
       )
       t2 <- Sys.time()
       t_fepoisson <- t2 - t1
-    }
-    t_fepoisson <- median(t_fepoisson)
+    # }
+    # t_fepoisson <- median(t_fepoisson)
 
     t_glm <- rep(0,10)
-    for (i in 1:0) {
+    # for (i in 1:0) {
       t1 <- Sys.time()
       fit <- suppressWarnings(glm(
         trade ~ log_dist + lang + cntg + clny + as.factor(exp_year) + as.factor(imp_year),
-        trade_panel_2006,
+        trade_panel_2,
         family = poisson(link = "log")
       ))
       t2 <- Sys.time()
       t_glm <- t2 - t1
-    }
-    t_glm <- median(t_glm)
+    # }
+    # t_glm <- median(t_glm)
 
     expect_lte(t_fepoisson, t_glm)
   }
