### 2024-08-28 7e0ea67
partial progress with srrstatsTODO
