### 2024-08-28 9ebfcba
minimal test for negbin
