### 2024-08-28 e51308b
additional ropensci requirements + minimal examples
