+++ src/cpp11.cpp
@@ -61,20 +61,6 @@ extern "C" SEXP _capybara_feglm_offset_fit_(SEXP eta_r, SEXP y_r, SEXP offset_r,
     return cpp11::as_sexp(feglm_offset_fit_(cpp11::as_cpp<cpp11::decay_t<const doubles &>>(eta_r), cpp11::as_cpp<cpp11::decay_t<const doubles &>>(y_r), cpp11::as_cpp<cpp11::decay_t<const doubles &>>(offset_r), cpp11::as_cpp<cpp11::decay_t<const doubles &>>(wt_r), cpp11::as_cpp<cpp11::decay_t<const std::string &>>(family), cpp11::as_cpp<cpp11::decay_t<const list &>>(control), cpp11::as_cpp<cpp11::decay_t<const list &>>(k_list)));
   END_CPP11
 }
-// 07_kendall_correlation.cpp
-double kendall_cor_(const doubles_matrix<> & m);
-extern "C" SEXP _capybara_kendall_cor_(SEXP m) {
-  BEGIN_CPP11
-    return cpp11::as_sexp(kendall_cor_(cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<> &>>(m)));
-  END_CPP11
-}
-// 07_kendall_correlation.cpp
-doubles pkendall_(doubles Q, int n);
-extern "C" SEXP _capybara_pkendall_(SEXP Q, SEXP n) {
-  BEGIN_CPP11
-    return cpp11::as_sexp(pkendall_(cpp11::as_cpp<cpp11::decay_t<doubles>>(Q), cpp11::as_cpp<cpp11::decay_t<int>>(n)));
-  END_CPP11
-}
 
 extern "C" {
 static const R_CallMethodDef CallEntries[] = {
@@ -86,8 +72,6 @@ static const R_CallMethodDef CallEntries[] = {
     {"_capybara_group_sums_cov_",      (DL_FUNC) &_capybara_group_sums_cov_,      3},
     {"_capybara_group_sums_spectral_", (DL_FUNC) &_capybara_group_sums_spectral_, 5},
     {"_capybara_group_sums_var_",      (DL_FUNC) &_capybara_group_sums_var_,      2},
-    {"_capybara_kendall_cor_",         (DL_FUNC) &_capybara_kendall_cor_,         1},
-    {"_capybara_pkendall_",            (DL_FUNC) &_capybara_pkendall_,            2},
     {NULL, NULL, 0}
 };
 }
