+++ R/fepoisson.R
@@ -6,16 +6,22 @@
 #' @inheritParams feglm
 #' 
 #' @examples
-#' # same as the example in feglm but with less typing
+#' # check the feglm examples for the details about clustered standard errors
+#' 
+#' # subset trade flows to avoid fitting time warnings during check
+#' set.seed(123)
+#' trade_2006 <- trade_panel[trade_panel$year == 2006, ]
+#' trade_2006 <- trade_2006[sample(nrow(trade_2006), 1000), ]
+#' 
 #' mod <- fepoisson(
 #'   trade ~ log_dist + lang + cntg + clny | exp_year + imp_year,
-#'   trade_panel
+#'   trade_2006
 #' )
 #'
 #' summary(mod)
 #' 
 #' @return A named list of class \code{"feglm"}.
-#' 
+#'
 #' @export
 fepoisson <- function(
     formula = NULL,
