+++ man/fepoisson.Rd
@@ -46,10 +46,16 @@ A wrapper for \code{\link{feglm}} with
 \code{family = poisson()}.
 }
 \examples{
-# same as the example in feglm but with less typing
+# check the feglm examples for the details about clustered standard errors
+
+# subset trade flows to avoid fitting time warnings during check
+set.seed(123)
+trade_2006 <- trade_panel[trade_panel$year == 2006, ]
+trade_2006 <- trade_2006[sample(nrow(trade_2006), 1000), ]
+
 mod <- fepoisson(
   trade ~ log_dist + lang + cntg + clny | exp_year + imp_year,
-  trade_panel
+  trade_2006
 )
 
 summary(mod)
