+++ man/fixed_effects.Rd
@@ -24,10 +24,16 @@ function has to be applied to our solution to get meaningful estimates of
 the fixed effects.
 }
 \examples{
-# same as the example in feglm but extracting the fixed effects
+# check the feglm examples for the details about clustered standard errors
+
+# subset trade flows to avoid fitting time warnings during check
+set.seed(123)
+trade_2006 <- trade_panel[trade_panel$year == 2006, ]
+trade_2006 <- trade_2006[sample(nrow(trade_2006), 1000), ]
+
 mod <- fepoisson(
   trade ~ log_dist + lang + cntg + clny | exp_year + imp_year,
-  trade_panel
+  trade_2006
 )
 
 fixed_effects(mod)
