### 2024-08-28 16beec6
simplify gh actions
