+++ R/generics_tidy.R
@@ -10,8 +10,14 @@ tidy.feglm <- function(x, conf.int = FALSE, conf.level = 0.95, ...) {
   res <- summary(x)$cm
   colnames(res) <- c("estimate", "std.error", "statistic", "p.value")
 
-  res[["conf.low"]] <- res[["estimate"]] - 1.96 * res[["std.error"]]
-  res[["conf.high"]] <- res[["estimate"]] + 1.96 * res[["std.error"]]
+  if (conf.int) {
+    res <- cbind(res, res[ , "estimate"] - 1.96 * res[ , "std.error"])
+    res <- cbind(res, res[ , "estimate"] + 1.96 * res[ , "std.error"])
+    colnames(res) <- c("estimate", "std.error", "statistic", "p.value",
+      "conf.low", "conf.high")
+  }
+
+  res <- as.data.frame(res)
 
   class(res) <- c("tbl_df", "tbl", "data.frame")
   res
