### 2024-09-06 b66ee27
add confint method
