### 2024-09-06 7dc4911
add confint method
