+++ NAMESPACE
@@ -1,12 +1,15 @@
 # Generated by roxygen2: do not edit by hand
 
 S3method(augment,feglm)
+S3method(augment,felm)
 S3method(coef,apes)
 S3method(coef,feglm)
 S3method(coef,felm)
 S3method(coef,summary.apes)
 S3method(coef,summary.feglm)
 S3method(coef,summary.felm)
+S3method(confint,feglm)
+S3method(confint,felm)
 S3method(fitted,feglm)
 S3method(fitted,felm)
 S3method(glance,feglm)
@@ -69,6 +72,7 @@ importFrom(stats,pnorm)
 importFrom(stats,poisson)
 importFrom(stats,predict)
 importFrom(stats,printCoefmat)
+importFrom(stats,qnorm)
 importFrom(stats,terms)
 importFrom(stats,vcov)
 importFrom(utils,combn)
