### 2024-09-06 #645 @maelle
@pachadotdev thanks for your submission! I'm fixing the content of your issue as you deleted HTML comments. Please do not delete them, they help the bot find the necessary metadata.
