### 2024-09-06 #645 @adamhsparks
Thanks, @maelle.

@pachadotdev, could you please address the issues flagged with 'srr'?
