### 2024-09-06 #645 @pachadotdev
> Thanks, @maelle.
>
> @pachadotdev, could you please address the issues flagged with 'srr'?

thanks @maelle and @adamhsparks

I shall check the ssr issues
