+++ R/felm.R
@@ -33,20 +33,20 @@
 #'
 #' @examples
 #' # check the feglm examples for the details about clustered standard errors
-#' 
+#'
 #' # subset trade flows to avoid fitting time warnings during check
 #' set.seed(123)
 #' trade_2006 <- trade_panel[trade_panel$year == 2006, ]
 #' trade_2006 <- trade_2006[sample(nrow(trade_2006), 1000), ]
 #'
-#' 
+#'
 #' mod <- felm(
 #'   log(trade) ~ log_dist + lang + cntg + clny | exp_year + imp_year,
 #'   trade_2006
 #' )
 #'
 #' summary(mod)
-#' 
+#'
 #' @export
 felm <- function(formula = NULL, data = NULL, weights = NULL) {
   # Use 'feglm' to estimate the model
