+++ tests/testthat/test-apes-bias.R
@@ -1,7 +1,7 @@
 test_that("apes/bias works", {
   trade_short <- trade_panel[trade_panel$year %in% 2002L:2006L, ]
   trade_short$trade <- ifelse(trade_short$trade > 100, 1L, 0L)
-  
+
   mod1 <- feglm(trade ~ lang | year, trade_short, family = binomial())
   apes1 <- apes(mod1)
   bias1 <- bias_corr(mod1)
