+++ tests/testthat/test-fepoisson.R
@@ -24,7 +24,7 @@ test_that("fepoisson is similar to fixest", {
 
   expect_output(print(mod))
 
-  expect_message(summary(mod))
+  # expect_message(summary(mod))
   expect_visible(summary(mod, type = "cluster"))
 
   fes <- fixed_effects(mod)
