+++ R/apes.R
@@ -93,14 +93,9 @@ apes <- function(
     sampling_fe = c("independence", "unrestricted"),
     weak_exo = FALSE) {
   # Check validity of 'object'
-  if (is.null(object)) {
-    stop("'object' has to be specified.", call. = FALSE)
-  } else if (!inherits(object, "feglm")) {
-    stop("'apes' called on a non-'feglm' object.", call. = FALSE)
-  }
+  apes_bias_check_object_(object, fun = "apes")
 
-  # Extract prior information if available or check validity of
-  # 'panel_structure'
+  # Extract prior information if available or check validity of panel_structure
   bias_corr <- inherits(object, "bias_corr")
   if (bias_corr) {
     panel_structure <- object[["panel_structure"]]
@@ -126,60 +121,20 @@ apes <- function(
   formula <- object[["formula"]]
   lvls_k <- object[["lvls_k"]]
   nt <- object[["nobs"]][["nobs"]]
-  nt.full <- object[["nobs"]][["nobs_full"]]
+  nt_full <- object[["nobs"]][["nobs_full"]]
   k <- length(lvls_k)
   k_vars <- names(lvls_k)
   p <- length(beta)
 
   # Check if binary choice model
-  if (family[["family"]] != "binomial") {
-    stop("'bias_corr' currently only supports binary choice models.",
-      call. = FALSE
-    )
-  }
+  apes_bias_check_binary_model_(family, fun = "apes")
 
   # Check if provided object matches requested panel structure
-  if (panel_structure == "classic") {
-    if (!(k %in% c(1L, 2L))) {
-      stop(
-        paste(
-          "panel_structure == 'classic' expects a one- or two-way fixed",
-          "effects model."
-        ),
-        call. = FALSE
-      )
-    }
-  } else {
-    if (!(k %in% c(2L, 3L))) {
-      stop(
-        paste(
-          "panel_structure == 'network' expects a two- or three-way fixed",
-          "effects model."
-        ),
-        call. = FALSE
-      )
-    }
-  }
+  apes_bias_check_panel_(panel_structure, k)
 
   # Check validity of 'n_pop'
   # Note: Default option is no adjustment i.e. only delta method covariance
-  if (!is.null(n_pop)) {
-    n_pop <- as.integer(n_pop)
-    if (n_pop < nt.full) {
-      warning(
-        paste(
-          "Size of the entire population is lower than the full sample size.",
-          "Correction factor set to zero."
-        ),
-        call. = FALSE
-      )
-      adj <- 0.0
-    } else {
-      adj <- (n_pop - nt.full) / (n_pop - 1L)
-    }
-  } else {
-    adj <- 0.0
-  }
+  adj <- apes_set_adj_(n_pop, nt_full)
 
   # Extract model response, regressor matrix, and weights
   y <- data[[1L]]
@@ -230,20 +185,20 @@ apes <- function(
       f1 <- family[["mu.eta"]](eta1)
       Delta[, j] <- (family[["linkinv"]](eta1) - family[["linkinv"]](eta0))
       Delta1[, j] <- f1 - family[["mu.eta"]](eta0)
-      J[, j] <- -colSums(PX * Delta1[, j]) / nt.full
-      J[j, j] <- sum(f1) / nt.full + J[j, j]
+      J[, j] <- -colSums(PX * Delta1[, j]) / nt_full
[…]
