+++ R/apes.R
@@ -66,7 +66,7 @@
 #' # subset trade flows to avoid fitting time warnings during check
 #' set.seed(123)
 #' trade_2006 <- trade_panel[trade_panel$year == 2006, ]
-#' trade_2006 <- trade_2006[sample(nrow(trade_2006), 1000), ]
+#' trade_2006 <- trade_2006[sample(nrow(trade_2006), 500), ]
 #'
 #' trade_2006$trade <- ifelse(trade_2006$trade > 100, 1L, 0L)
 #'
@@ -99,8 +99,8 @@ apes <- function(
   bias_corr <- inherits(object, "bias_corr")
   if (bias_corr) {
     panel_structure <- object[["panel_structure"]]
-    L <- object[["bandwidth"]]
-    if (L > 0L) {
+    l <- object[["bandwidth"]]
+    if (l > 0L) {
       weak_exo <- TRUE
     } else {
       weak_exo <- FALSE
@@ -116,7 +116,6 @@ apes <- function(
   beta <- object[["coefficients"]]
   control <- object[["control"]]
   data <- object[["data"]]
-  eps <- .Machine[["double.eps"]]
   family <- object[["family"]]
   formula <- object[["formula"]]
   lvls_k <- object[["lvls_k"]]
@@ -139,7 +138,7 @@ apes <- function(
   # Extract model response, regressor matrix, and weights
   y <- data[[1L]]
   x <- model.matrix(formula, data, rhs = 1L)[, -1L, drop = FALSE]
-  nms.sp <- attr(x, "dimnames")[[2L]]
+  nms_sp <- attr(x, "dimnames")[[2L]]
   attr(x, "dimnames") <- NULL
   wt <- object[["weights"]]
 
@@ -152,12 +151,12 @@ apes <- function(
   # Compute derivatives and weights
   eta <- object[["eta"]]
   mu <- family[["linkinv"]](eta)
-  mu.eta <- family[["mu.eta"]](eta)
+  mu_eta <- family[["mu.eta"]](eta)
   v <- wt * (y - mu)
-  w <- wt * mu.eta
+  w <- wt * mu_eta
   z <- wt * partial_mu_eta_(eta, family, 2L)
   if (family[["link"]] != "logit") {
-    h <- mu.eta / family[["variance"]](mu)
+    h <- mu_eta / family[["variance"]](mu)
     v <- h * v
     w <- h * w
     z <- h * z
@@ -172,68 +171,68 @@ apes <- function(
   }
 
   # Compute average partial effects, derivatives, and Jacobian
-  PX <- x - mx
-  Delta <- matrix(NA_real_, nt, p)
-  Delta1 <- matrix(NA_real_, nt, p)
-  J <- matrix(NA_real_, p, p)
-  Delta[, !binary] <- mu.eta
-  Delta1[, !binary] <- partial_mu_eta_(eta, family, 2L)
-  for (j in seq.int(p)) {
-    if (binary[[j]]) {
-      eta0 <- eta - x[, j] * beta[[j]]
-      eta1 <- eta0 + beta[[j]]
+  px <- x - mx
+  delta <- matrix(NA_real_, nt, p)
+  delta1 <- matrix(NA_real_, nt, p)
+  j <- matrix(NA_real_, p, p)
+  delta[, !binary] <- mu_eta
+  delta1[, !binary] <- partial_mu_eta_(eta, family, 2L)
+  for (i in seq.int(p)) {
+    if (binary[[i]]) {
+      eta0 <- eta - x[, i] * beta[[i]]
+      eta1 <- eta0 + beta[[i]]
       f1 <- family[["mu.eta"]](eta1)
[…]
