### 2024-09-08 9b351e7
consistent generics
