### 2024-09-08 #645 @adamhsparks
> hi @adamhsparks I fixed the check issues
>
> The only thing I am not sure about is this: ✖️ All applicable standards [v0.2.0] have been documented in this package (99 complied with; 17 N/A standards) Statistical standards should be documented in most package files, yet are mostly only documented in one file.
>
> I tried different combinations and srr always creates a single file with the review tasks

@mpadge, can you provide any insight here?
