### 2024-09-10 #645 @mpadge
@pachadotdev [Our *Stats Devguide* refers to this "single file" and states](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-srr):

> The first task after having generated this file is to move standards to approximate locations within your package where they are likely to be addressed. For example, standards concerning tests should be moved somewhere within the tests/ directory, standards concerning documentation to the main README.Rmd file, or within a vignette file...
>
> Moving different standards to more appropriate locations within your code will break down an initially large single list of standards into more manageable groups dispersed throughout your code. As each standard is addressed, it should be moved to one or more locations in your code as near as possible to relevant code, the tag changed from @srrstatsTODO to @srrstats, and a brief description appended to explain how that standard is addressed.

See the [main `srr` vignette](https://ropensci-review-tools.github.io/srr/articles/srr-stats.html) referred to in the *Stats Devguide* for more detail. For each individual standard, both general and category-specific, you need to describe how your code complies with the standard, *and move the standard itself* to the point or points within the code where that standard applies. Note that code will commonly adhere to single standards at multiple points, and that multiple copies of `@srrstats` tags are generally expected.

In short: The "srr-stats-standards.R" that is auto-generated by the `srr` package is just a template for you to use in *starting* to document how your code complies with each standard. From there, you have to work through each one, and modify the text to explain how you adhere to the standard.
