### 2024-10-17 3f2cb7b
reorganize srr comments
