+++ R/generics_vcov.R
@@ -1,0 +1,239 @@
+#' @title Covariance matrix for APEs
+#'
+#' @description Covariance matrix for the estimator of the
+#'  average partial effects from objects returned by \code{\link{apes}}.
+#'
+#' @srrstats {RE4.6} *The variance-covariance matrix of the model parameters (via `vcov()`)*
+#' 
+#' @param object an object of class \code{"apes"}.
+#' @param ... additional arguments.
+#'
+#' @return A named matrix of covariance estimates.
+#'
+#' @seealso \code{\link{apes}}
+#'
+#' @export
+#'
+#' @noRd
+vcov.apes <- function(object, ...) {
+  object[["vcov"]]
+}
+
+#' @title Covariance matrix for GLMs
+#'
+#' @description Covariance matrix for the estimator of the structural parameters
+#'  from objects returned by \code{\link{feglm}}. The covariance is computed
+#' from the hessian, the scores, or a combination of both after convergence.
+#'
+#' @srrstats {RE4.6} *The variance-covariance matrix of the model parameters (via `vcov()`)*
+#' 
+#' @param object an object of class \code{"feglm"}.
+#' @param type the type of covariance estimate required. \code{"hessian"} refers
+#'  to the inverse of the negative expected hessian after convergence and is the
+#'  default option. \code{"outer.product"} is the outer-product-of-the-gradient
+#'  estimator. \code{"sandwich"} is the sandwich estimator (sometimes also
+#'  referred as robust estimator), and \code{"clustered"} computes a clustered
+#'  covariance matrix given some cluster variables.
+#'
+#' @param ... additional arguments.
+#'
+#' @return A named matrix of covariance estimates.
+#'
+#' @references Cameron, C., J. Gelbach, and D. Miller (2011). "Robust Inference
+#'  With Multiway Clustering". Journal of Business & Economic Statistics 29(2).
+#'
+#' @seealso \code{\link{feglm}}
+#'
+#' @examples
+#' # same as the example in feglm but extracting the covariance matrix
+#'
+#' # subset trade flows to avoid fitting time warnings during check
+#' set.seed(123)
+#' trade_2006 <- trade_panel[trade_panel$year == 2006, ]
+#' trade_2006 <- trade_2006[sample(nrow(trade_2006), 500), ]
+#'
+#' mod <- fepoisson(
+#'   trade ~ log_dist + lang + cntg + clny | exp_year + imp_year | pair,
+#'   trade_2006
+#' )
+#'
+#' round(vcov(mod, type = "clustered"), 5)
+#'
+#' @return A named matrix of covariance estimates.
+#'
+#' @export
+vcov.feglm <- function(
+    object,
+    type = c("hessian", "outer.product", "sandwich", "clustered"),
+    ...) {
+  # Check validity of input argument 'type'
+  type <- match.arg(type)
+
+  # Extract cluster from formula
+  # it is totally fine not to have a cluster variable
+  cl_vars <- vcov_feglm_vars_(object)
+  k <- length(cl_vars)
+
+  if (isTRUE(k >= 1L) && type != "clustered") {
+    type <- "clustered"
+  }
+
+  # Compute requested type of covariance matrix
+  h <- object[["hessian"]]
+  p <- ncol(h)
+
+  if (type == "hessian") {
+    # If the hessian is invertible, compute its inverse
+    v <- vcov_feglm_hessian_covariance_(h, p)
+  } else {
+    g <- get_score_matrix_(object)
+    if (type == "outer.product") {
[…]
