+++ R/generics_augment.R
@@ -4,6 +4,8 @@ generics::augment
 
 #' @title Broom Integration
 #'
+#' @srrstats {RE4.10} *Model Residuals, including sufficient documentation to enable interpretation of residuals, and to enable users to submit residuals to their own tests.*
+#' 
 #' @description The provided `broom` methods do the following:
 #'  1. `augment`: takes the input data and adds additional columns with the
 #'      fitted values and residuals.
