+++ man/felm.Rd
@@ -51,7 +51,6 @@ set.seed(123)
 trade_2006 <- trade_panel[trade_panel$year == 2006, ]
 trade_2006 <- trade_2006[sample(nrow(trade_2006), 500), ]
 
-
 mod <- felm(
   log(trade) ~ log_dist + lang + cntg + clny | exp_year + imp_year,
   trade_2006

[…]
