### 2024-10-17 #645 @pachadotdev
thanks @mpadge
I re-organized the srr comments. The general aspects are still in the main srr file, while the specifics were moved to different scripts and therefore sometimes there is a duplication of the same tag (e.g., such as a `vcov` method in fileA.r and fileB.r)
