### 2024-10-23 #645 @mpadge
@pachadotdev You can call `check srr` yourself at any time. The above output shows what you still need to do to document compliance. Please read the [main srr guide](https://docs.ropensci.org/srr/articles/srr-stats.html), and adhere to all requirements there. This includes documenting compliance in `/tests` files, as well as [standards for documentation](https://docs.ropensci.org/srr/articles/srr-stats.html#documenting-standards-for-documentation), both in `README.Rmd`, and in vignettes. Once you've done that, please use [`srr::srr_stats_pre_submit()`](https://docs.ropensci.org/srr/reference/srr_stats_pre_submit.html) locally to confirm compliance. Only once that indicates that everything is okay, feel free to call `check srr` here again to let us know. Thanks
