### 2024-10-25 05b7c14
add srr documentation in tests/
