+++ tests/testthat/test-fepoisson.R
@@ -1,3 +1,11 @@
+#' srr_stats (tests)
+#'
+#' @srrstatsVerbose TRUE
+#' 
+#' @srrstats {G5.12} *Any conditions necessary to run extended tests such as platform requirements, memory, expected runtime, and artefacts produced that may need manual inspection, should be described in developer documentation such as a `CONTRIBUTING.md` or `tests/README.md` file.*
+#' @noRd 
+NULL
+
 test_that("fepoisson is similar to fixest", {
   mod <- fepoisson(
     trade ~ log_dist + lang + cntg + clny | exp_year + imp_year | pair,
@@ -79,3 +87,41 @@ test_that("fepoisson is similar to fixest", {
     expect_lte(t_fepoisson, t_glm)
   }
 })
+
+# test_that("fepoisson time is the same adding noise to the data", {
+#   trade_panel2 <- trade_panel
+#   set.seed(200100)
+#   trade_panel2$trade2 <- trade_panel$trade + rbinom(nrow(trade_panel2), 1, 0.5) *
+#     .Machine$double.eps
+#   m1 <- fepoisson(
+#     trade ~ log_dist + lang + cntg + clny | exp_year + imp_year | pair,
+#     trade_panel2
+#   )
+#   m2 <- fepoisson(
+#     trade2 ~ log_dist + lang + cntg + clny | exp_year + imp_year | pair,
+#     trade_panel2
+#   )
+#   expect_equal(coef(m1), coef(m2))
+#   expect_equal(fixed_effects(m1), fixed_effects(m2))
+
+#   t1 <- rep(NA, 10)
+#   t2 <- rep(NA, 10)
+#   for (i in 1:10) {
+#     a <- Sys.time()
+#     m1 <- fepoisson(
+#       trade ~ log_dist + lang + cntg + clny | exp_year + imp_year | pair,
+#       trade_panel2
+#     )
+#     b <- Sys.time()
+#     t1[i] <- b - a
+
+#     a <- Sys.time()
+#     m2 <- fepoisson(
+#       trade2 ~ log_dist + lang + cntg + clny | exp_year + imp_year | pair,
+#       trade_panel2
+#     )
+#     b <- Sys.time()
+#     t2[i] <- b - a
+#   }
+#   expect_lte(abs(median(t1) - median(t2)), 0.05)
+# })
