+++ tests/testthat/test-fenegbin.R
@@ -3,6 +3,7 @@
 #' @srrstatsVerbose TRUE
 #'
 #' @srrstats {G5.12} *Any conditions necessary to run extended tests such as platform requirements, memory, expected runtime, and artefacts produced that may need manual inspection, should be described in developer documentation such as a `CONTRIBUTING.md` or `tests/README.md` file.*
+#' 
 #' @noRd
 NULL
 
