### 2024-10-25 #645 @pachadotdev
sorry @mpadge , I must assume my dumminess now

I tried

```
> srr::srr_stats_pre_submit()
✔ All applicable standards [v0.2.0] have been documented in this package (190 complied with; 18 N/A standards)
! Standards should be documented in multiple directories, yet are only present in one
```

Even when I documented SRR standards here https://github.com/pachadotdev/capybara/blob/main/tests/testthat/test-feglm.R and in other tests, it still says that the tests are only in one directory
