### 2024-10-27 #645 @pachadotdev
> Hmmm, and the ❗ is the one you mean, yet the text has been dropped. I'll check it out for you asap and report back. Thanks

it seems to the the srr pkg itself, let me test
