### 2024-10-27 #645 @pachadotdev
@mpadge the problem in srr is that it uses `base_path = "."` that combined with `gsub` goes into "Hulk" mode and removes all the paths

I proposed a fix here https://github.com/ropensci-review-tools/srr/pull/45/files#diff-c807a8e5a9d16e9a6c853c660ac8108f799a50d9a7c6f6c99ade0665f7caa28aR92
