### 2024-10-28 #645 @mpadge
@pachadotdev thanks, but that wasn't the issue; I think recent updates just hadn't yet been incorporated in the bot deployment. Everything seems fine, and i see you've updated the stats documentation, so feel free to call `check srr` again at any time.
