### 2024-11-17 pachadotdev/capybara#11 @pachadotdev
finish rOpenSci standard documentation
https://github.com/ropensci/software-review/issues/645#issuecomment-2441095565
