### 2024-11-17 2a8d78e
apply proper ssr documentation
