### 2024-11-20 dac82ff
expand predict method + distributed srr internal documentation
