### 2024-11-20 1ad5deb
add configure + missing srr documentation entries
