+++ tests/testthat/test-errors.R
@@ -1,5 +1,7 @@
 #' srr_stats (tests)
 #' @srrstatsVerbose TRUE
+#' @srrstats {G5.2} *Appropriate error and warning behaviour of all functions
+#'  should be explicitly demonstrated through tests. In particular,*
 #' @srrstats {G5.2b} Here we cover errors as broadly as possible conditions,
 #'  and we write intentionally bad examples to compare the result with expected
 #'  values and inputs.
