### 2024-11-20 #645 @pachadotdev
@mpadge Hi! I spreaded the documentation as you indicated. For most items I used the same wording from srr, because those were already clear. For some cases I explained with a new writing when it was pertinent for clarity.
