### 2024-11-22 4e59d5e
I explained each srr section in my own words
