+++ R/srr-stats-standards.R
@@ -1,68 +1,0 @@
-#' srr_stats
-#'
-#' @srrstatsVerbose TRUE
-#'
-#' @srrstats {G1.5} Software should include all code necessary to reproduce results which form the basis of performance claims made in associated publications.
-#' @srrstats {G1.6} Software should include code necessary to compare performance claims with alternative implementations in other R packages.
-#' @srrstats {G2.0a} Provide explicit secondary documentation of any expectations on lengths of inputs
-#' @srrstats {G2.1} Implement assertions on types of inputs (see the initial point on nomenclature above).
-#' @srrstats {G2.1a} Provide explicit secondary documentation of expectations on data types of all vector inputs.
-#' @srrstats {G2.2} Appropriately prohibit or restrict submission of multivariate input to parameters expected to be univariate.
-#' @srrstats {G2.4} Provide appropriate mechanisms to convert between different data types, potentially including:
-#' @srrstats {G2.4a} explicit conversion to `integer` via `as.integer()`
-#' @srrstats {G2.4b} explicit conversion to continuous via `as.numeric()`
-#' @srrstats {G2.4c} explicit conversion to character via `as.character()` (and not `paste` or `paste0`)
-#' @srrstats {G2.4d} explicit conversion to factor via `as.factor()`
-#' @srrstats {G2.4e} explicit conversion from factor via `as...()` functions
-#' @srrstats {G2.5} Where inputs are expected to be of `factor` type, secondary documentation should explicitly state whether these should be `ordered` or not, and those inputs should provide appropriate error or other routines to ensure inputs follow these expectations.
-#' @srrstats {G2.6} Software which accepts one-dimensional input should ensure values are appropriately pre-processed regardless of class structures.
-#' @srrstats {G2.7} Software should accept as input as many of the above standard tabular forms as possible, including extension to domain-specific forms.
-#' @srrstats {G2.8} Software should provide appropriate conversion or dispatch routines as part of initial pre-processing to ensure that all other sub-functions of a package receive inputs of a single defined class or type.
-#' @srrstats {G2.9} Software should issue diagnostic messages for type conversion in which information is lost (such as conversion of variables from factor to character; standardisation of variable names; or removal of meta-data such as those associated with [`sf`-format](https://r-spatial.github.io/sf/) data) or added (such as insertion of variable or column names where none were provided).
-#' @srrstats {G2.10} Software should ensure that extraction or filtering of single columns from tabular inputs should not presume any particular default behaviour, and should ensure all column-extraction operations behave consistently regardless of the class of tabular data used as input.
-#' @srrstats {G2.11} Software should ensure that `data.frame`-like tabular objects which have columns which do not themselves have standard class attributes (typically, `vector`) are appropriately processed, and do not error without reason. This behaviour should be tested. Again, columns created by the [`units` package](https://github.com/r-quantities/units/) provide a good test case.
-#' @srrstats {G2.12} Software should ensure that `data.frame`-like tabular objects which have list columns should ensure that those columns are appropriately pre-processed either through being removed, converted to equivalent vector columns where appropriate, or some other appropriate treatment such as an informative error. This behaviour should be tested.
-#' @srrstats {G2.13} Statistical Software should implement appropriate checks for missing data as part of initial pre-processing prior to passing data to analytic algorithms.
-#' @srrstats {G2.14} Where possible, all functions should provide options for users to specify how to handle missing (`NA`) data, with options minimally including:
-#' @srrstats {G2.14a} error on missing data
-#' @srrstats {G2.14b} ignore missing data with default warnings or messages issued
[…]
