### 2024-11-23 af36b88
document all standards
