+++ R/bias_corr.R
@@ -3,11 +3,16 @@
 #' @srrstats {G2.1a} Ensures input objects are of the expected class (`feglm`).
 #' @srrstats {G2.3a} Validates string arguments like `panel_structure` using `match.arg()` for predefined values.
 #' @srrstats {G2.14a} Provides errors for missing or invalid inputs, such as non-`feglm` objects.
-#' @srrstats {G2.14b} Provides clear error messages when exceeding supported fixed-effect dimensions.
+#' @srrstats {G2.14b} Provides clear error messages when the data structure is incompatible with the model requirements.
 #' @srrstats {G3.1a} Supports structured panels (`classic` or `network`) for analyzing fixed effects.
 #' @srrstats {RE5.0} Efficient handling of computational scaling for large panels through fixed-effect groupings.
 #' @srrstats {G5.2a} Produces unique and informative error messages for all stopping conditions.
 #' @srrstats {G5.4a} Includes logical checks for computational edge cases, such as unsupported models.
+#' @noRd
+NULL
+
+#' NA_standards
+#' @srrstatsNA {G2.14} Missing observations are dropped, otherwise providing imputation methods would bias the estimation (i.e., replacing all missing values with the median).
 #' @noRd
 NULL
 

[…]
