### 2024-11-23 5535f88
add standard 7.1a
