### 2024-11-23 #645 @pachadotdev
hi @mpadge
I took a different approach now, I and proceeded to explain the code in the srr headers. Some explanations are too obvious but I did exactly what was requested. My approach was "not everybody has read Hansen's Econometrics, so it is not obvious how to pre-process the data"
