### 2024-12-01 #645 @adamhsparks
Hi, @pachadotdev, my time as EIC ended at the beginning of last month. Thank you for all your work to get {Capybara} to this point.

Pinging, @emilyriederer as the current acting EIC to action your submission.
