### 2024-12-10 #645 @emilyriederer
Hi @pachadotdev ! Our general guidance is that authors should have one live review at a time to manage overall capacity. Do you want to move forward here and we'll put `{redatam}` on hold for now?
