### 2025-02-02 #645 @maurolepore
Dear @pachadotdev and @rkillick just to let you know that my EiC rotation just started and I'm checking all open issues.

Looks like the package passes all checks cleanly 💯

@rkillick are you still able to handle this submission?
