### 2025-02-28 #645 @pachadotdev
> [@maurolepore](https://github.com/maurolepore) Yes, I've been trying to find reviewers.

hi @rkillick, can I send the package to CRAN in the meanwhile?
