### 2025-02-28 #645 @maurolepore
@pachadotdev while I'm here I'll share the relevant guideline:

> We strongly suggest submitting your package for review before publishing on CRAN or submitting a software paper describing the package to a journal. Review feedback may result in major improvements and updates to your package, including renaming and breaking changes to functions.
-- https://devdevguide.netlify.app/softwarereview_author

So I interpret this as we don't encourage it but also we won't stop you.
