### 2025-05-04 #645 @maurolepore
@pachadotdev I'm wrapping up my EiC rotation but before I hand on this submission to the next EiC I wanted to thank you for all your contributions to rOpenSci.
