### 2026-01-31 #645 @emilyriederer
Hi @rkillick and @pachadotdev -- I have just begun my EiC rotation and am checking in on stalled reviews. Are we still actively seeking reviewers for this submission?

As a note mostly to myself, this open review was blocking beginning work on #672 . So, we shall revisit that issue once we clarify the path forward here.
